+++ FM.py
@@ -449,6 +449,16 @@ TBGLST = [
 	("FMAXPSG_SCTN0905_TUPDICTSTRSTRADD", FMAXFM_SCTN0101_AXDEF, "add a str to a tupdict <NAC><TUPDICTNAME><STR><STR>",),
 	("FMAXPSG_SCTN0905_TUPDICTSTRVALADD", FMAXFM_SCTN0101_AXDEF, "add a value to a tupdict <NAC><TUPDICTNAME><STR><VAL>",),
 	("FMAXPSG_SCTN0905_TUPDICTVALADD", FMAXFM_SCTN0101_AXDEF, "add a value to a tupdict <NAC><TUPDICTNAME><KEY><VAL>",),
+	("FMAXPSG_SCTN0906_BTNDEF", FMAXFM_SCTN0101_AXDEF, "define a button <NAC><BTNNAME>",),
+	("FMAXPSG_SCTN0906_BTN_TUPVALVALADD", FMAXFM_SCTN0101_AXDEF, "add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL1><VAL2>",),
+	("FMAXPSG_SCTN0906_BTN_VALADD", FMAXFM_SCTN0101_AXDEF, "add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL>",),
+	("FMAXPSG_SCTN0906_BTN_STRADD", FMAXFM_SCTN0101_AXDEF, "add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL>",),
+	("FMAXPSG_SCTN0907_SPINDEF", FMAXFM_SCTN0101_AXDEF, "define a spin box entry <NAC><SPINNAME>",),
+	("FMAXPSG_SCTN0907_SPIN_VALUESSTRADD", FMAXFM_SCTN0101_AXDEF, "add a STR to the values list <NAC><SPINNAME><STR>",),
+	("FMAXPSG_SCTN0907_SPIN_VALUESVALADD", FMAXFM_SCTN0101_AXDEF, "add a VAL to the values list <NAC><SPINNAME><VAL>",),
+	("FMAXPSG_SCTN0907_SPIN_DICTSTRADD", FMAXFM_SCTN0101_AXDEF, "add a STR to the SPINDICT <NAC><SPINNAME><KEY><STR>",),
+	("FMAXPSG_SCTN0907_SPIN_DICTTUPVALVALADD", FMAXFM_SCTN0101_AXDEF, "add a (VAL1, VAL2) to the SPINDICT <NAC><SPINNAME><KEY><VAL1><VAL2>",),
+	("FMAXPSG_SCTN0907_SPIN_DICTSTRADD", FMAXFM_SCTN0101_AXDEF, "add a VAL to the SPINDICT <NAC><SPINNAME><KEY><VAL>",),
 	("FMAXPSG_____", FMAX_NOP, "FMAXPSG_ENDS",),
 	("FMAX_NOP", FMAXFM_SCTN0101_AXDEF, "skip this entry",),
 	("FMCF", FMAX_NOP, "FMCF_BEGINS",),
@@ -495,103 +505,124 @@ TBGLST = [
 	("FMPSG_SCTN0905_TUPDICTDICT", FMAXFM_SCTN0103_DICTDEF, "define the dict to hold everything in SCTN0900",),
 	("FMPSG_____", FMAX_NOP, "FMPSG_ENDS",),
 	("PSGVAL", FMAX_NOP, "FMPSG_BEGINS",),
-	("PSGVAL__BUTTON", FMAXPSG_SCTN0905_TUPDICTDEF, "BUTTON", "define the button empty tupdict",),
-	("PSGVAL_BUTTON_AUTO_SIZE_BUTTON", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "AUTO_SIZE_BUTTON", "None", "",),
-	("PSGVAL_BUTTON_BIND_RETURN_KEY", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BIND_RETURN_KEY", "False", "",),
-	("PSGVAL_BUTTON_BORDER_WIDTH", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BORDER_WIDTH", "None", "",),
-	("PSGVAL_BUTTON_BUTTON_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BUTTON_COLOR", "None", "",),
-	("PSGVAL_BUTTON_BUTTON_TEXT", FMAXPSG_SCTN0905_TUPDICTSTRADD, "BUTTON", "BUTTON_TEXT", "", "",),
-	("PSGVAL_BUTTON_BUTTON_TYPE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BUTTON_TYPE", "7", "",),
-	("PSGVAL_BUTTON_CHANGE_SUBMITS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "CHANGE_SUBMITS", "False", "",),
-	("PSGVAL_BUTTON_DEFAULT_EXTENSION", FMAXPSG_SCTN0905_TUPDICTSTRADD, "BUTTON", "DEFAULT_EXTENSION", "", "",),
-	("PSGVAL_BUTTON_DISABLED", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "DISABLED", "False", "",),
-	("PSGVAL_BUTTON_DISABLED_BUTTON_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "DISABLED_BUTTON_COLOR", "None", "",),
-	("PSGVAL_BUTTON_ENABLE_EVENTS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "ENABLE_EVENTS", "False", "",),
-	("PSGVAL_BUTTON_FILE_TYPES", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "FILE_TYPES", "(('ALL FILES', '*.*'),)", "",),
-	("PSGVAL_BUTTON_FOCUS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "FOCUS", "False", "",),
-	("PSGVAL_BUTTON_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "FONT", "None", "",),
-	("PSGVAL_BUTTON_HIGHLIGHT_COLORS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "HIGHLIGHT_COLORS", "None", "",),
-	("PSGVAL_BUTTON_IMAGE_DATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_DATA", "None", "",),
-	("PSGVAL_BUTTON_IMAGE_FILENAME", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_FILENAME", "None", "",),
-	("PSGVAL_BUTTON_IMAGE_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_SIZE", "(None, None)", "",),
-	("PSGVAL_BUTTON_IMAGE_SUBSAMPLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_SUBSAMPLE", "None", "",),
-	("PSGVAL_BUTTON_INITIAL_FOLDER", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "INITIAL_FOLDER", "None", "",),
-	("PSGVAL_BUTTON_K", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "K", "None", "",),
-	("PSGVAL_BUTTON_KEY", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "KEY", "None", "",),
-	("PSGVAL_BUTTON_METADATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "METADATA", "None", "",),
-	("PSGVAL_BUTTON_PAD", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "PAD", "None", "",),
-	("PSGVAL_BUTTON_RIGHT_CLICK_MENU", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "RIGHT_CLICK_MENU", "None", "",),
-	("PSGVAL_BUTTON_S", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "S", "(None, None)", "",),
-	("PSGVAL_BUTTON_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "SIZE", "(None, None)", "",),
-	("PSGVAL_BUTTON_TARGET", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "TARGET", "(None, None))", "",),
-	("PSGVAL_BUTTON_TOOLTIP", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "TOOLTIP", "None", "",),
-	("PSGVAL_BUTTON_USE_TTK_BUTTONS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "USE_TTK_BUTTONS", "None", "",),
-	("PSGVAL_BUTTON_VISIBLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "VISIBLE", "True", "",),
-	("PSGVAL_COLOR_BACKGROUND", FMAXPSG_SCTN0900_STRDEF, "BACKGROUNDCOLOR", "#330022", "the background of the main frames",),
+	("PSGVAL_BUTTON", FMAXPSG_SCTN0905_TUPDICTDEF, "BUTTON", "define the button empty tupdict",),
+	("PSGVAL_BUTTON_AUTO_SIZE_BUTTON", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "AUTO_SIZE_BUTTON", "None", "if True the button size is sized to fit the text",),
+	("PSGVAL_BUTTON_BIND_RETURN_KEY", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BIND_RETURN_KEY", "False", "If True the return key will cause this button to be pressed",),
+	("PSGVAL_BUTTON_BORDER_WIDTH", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BORDER_WIDTH", "None", "width of border around button in pixels",),
+	("PSGVAL_BUTTON_BUTTON_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BUTTON_COLOR", "None", "Color of button. default is from theme or the window. Easy to remember which is which if you say 'ON' between colors. 'red' on 'green'. Normally a tuple, but can be a simplified-button-color-string 'foreground on background'. Can be a single color if want to set only the background.",),
+	("PSGVAL_BUTTON_BUTTON_TEXT", FMAXPSG_SCTN0905_TUPDICTSTRADD, "BUTTON", "BUTTON_TEXT", "", "str text to display on the button",),
+	("PSGVAL_BUTTON_BUTTON_TYPE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "BUTTON_TYPE", "7", "You  should NOT be setting this directly. ONLY the shortcut functions set this",),
+	("PSGVAL_BUTTON_CHANGE_SUBMITS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "CHANGE_SUBMITS", "False", "DO NOT USE. Only listed for backwards compat - Use enable_events instead",),
+	("PSGVAL_BUTTON_DEFAULT_EXTENSION", FMAXPSG_SCTN0905_TUPDICTSTRADD, "BUTTON", "DEFAULT_EXTENSION", "", "If no extension entered by user, add this to filename (only used in saveas dialogs)",),
+	("PSGVAL_BUTTON_DISABLED", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "DISABLED", "False", "If True button will be created disabled. If BUTTON_DISABLED_MEANS_IGNORE then the button will be ignored rather than disabled using tkinter",),
+	("PSGVAL_BUTTON_DISABLED_BUTTON_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "DISABLED_BUTTON_COLOR", "None", "colors to use when button is disabled (text, background). Use None for a color if don't want to change. Only ttk buttons support both text and background colors. tk buttons only support changing text color",),
+	("PSGVAL_BUTTON_ENABLE_EVENTS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "ENABLE_EVENTS", "False", "Turns on the element specific events. If this button is a target, should it generate an event when filled in",),
+	("PSGVAL_BUTTON_FILE_TYPES", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "FILE_TYPES", "(('ALL FILES', '*.*'),)", "the filetypes that will be used to match files. To indicate all files: (('ALL Files', '*.*'),).  Note - NOT SUPPORTED ON MAC",),
+	("PSGVAL_BUTTON_FOCUS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "FOCUS", "False", "if True, initial focus will be put on this button",),
+	("PSGVAL_BUTTON_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "FONT", "None", "specifies the font family, size, etc",),
+	("PSGVAL_BUTTON_HIGHLIGHT_COLORS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "HIGHLIGHT_COLORS", "None", "colors to use when button has focus (highlight, background). None will use computed colors. Only used by Linux and only for non-TTK button",),
+	("PSGVAL_BUTTON_IMAGE_DATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_DATA", "None", "Raw or Base64 representation of the image to put on button. Choose either filename or data",),
+	("PSGVAL_BUTTON_IMAGE_FILENAME", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_FILENAME", "None", "image filename if there is a button image. GIFs and PNGs only.",),
+	("PSGVAL_BUTTON_IMAGE_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_SIZE", "(None, None)", "Size of the image in pixels (width, height)",),
+	("PSGVAL_BUTTON_IMAGE_SUBSAMPLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "IMAGE_SUBSAMPLE", "None", "amount to reduce the size of the image. Divides the size by this number. 2=1/2, 3=1/3, 4=1/4, etc",),
+	("PSGVAL_BUTTON_INITIAL_FOLDER", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "INITIAL_FOLDER", "None", "starting path for folders and files",),
+	("PSGVAL_BUTTON_K", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "K", "None", "Used with window.FindElement and with return values to uniquely identify this element to uniquely identify this element",),
+	("PSGVAL_BUTTON_KEY", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "KEY", "None", "Used with window.FindElement and with return values to uniquely identify this element to uniquely identify this element",),
+	("PSGVAL_BUTTON_METADATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "METADATA", "None", "User metadata that can be set to ANYTHING",),
+	("PSGVAL_BUTTON_PAD", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "PAD", "None", "Amount of padding to put around element (left/right, top/bottom) or ((left, right), (top, bottom))",),
+	("PSGVAL_BUTTON_RIGHT_CLICK_MENU", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "RIGHT_CLICK_MENU", "None", "A list of lists of Menu items to show when this element is right clicked. See user docs for exact format.",),
+	("PSGVAL_BUTTON_S", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "S", "(None, None)", "(width, height) of the button in characters wide, rows high",),
+	("PSGVAL_BUTTON_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "SIZE", "(None, None)", "(width, height) of the button in characters wide, rows high",),
+	("PSGVAL_BUTTON_TARGET", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "TARGET", "(None, None)", "str | Tuple[int, int] key or (row,col) target for the button. Note that -1 for column means 1 element to the left of this one. The constant ThisRow is used to indicate the current row. The Button itself is a valid target for some types of button",),
+	("PSGVAL_BUTTON_TOOLTIP", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "TOOLTIP", "None", "text, that will appear when mouse hovers over the element",),
+	("PSGVAL_BUTTON_USE_TTK_BUTTONS", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "USE_TTK_BUTTONS", "None", "True = use ttk buttons. False = do not use ttk buttons.  None (Default) = use ttk buttons only if on a Mac and not with button images",),
+	("PSGVAL_BUTTON_VISIBLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "BUTTON", "VISIBLE", "True", "set visibility state of the element",),
+	("PSGVAL_CHECKBOX", FMAXPSG_SCTN0905_TUPDICTDEF, "CHECKBOX", "set a checkbox in motion",),
+	("PSGVAL_CHECKBOX_AUTO_SIZE_TEXT", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "AUTO_SIZE_TEXT", "None", "if True will size the element to match the length of the text",),
+	("PSGVAL_CHECKBOX_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "BACKGROUND_COLOR", "None", "color of background",),
+	("PSGVAL_CHECKBOX_CHANGE_SUBMITS", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "CHANGE_SUBMITS", "False", "DO NOT USE. Only listed for backwards compat - Use enable_events instead",),
+	("PSGVAL_CHECKBOX_CHECKBOX_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "CHECKBOX_COLOR", "None", "color of background of the box that has the check mark in it. The checkmark is the same color as the text",),
+	("PSGVAL_CHECKBOX_DEFAULT", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "DEFAULT", "False", "Set to True if you want this checkbox initially checked",),
+	("PSGVAL_CHECKBOX_DISABLED", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "DISABLED", "False", "set disable state",),
+	("PSGVAL_CHECKBOX_ENABLE_EVENTS", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "ENABLE_EVENTS", "False", "Turns on the element specific events. Checkbox events happen when an item changes",),
+	("PSGVAL_CHECKBOX_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "FONT", "None", "specifies the font family, size, etc",),
+	("PSGVAL_CHECKBOX_K", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "K", "None", "Used with window.FindElement and with return values to uniquely identify this element",),
+	("PSGVAL_CHECKBOX_KEY", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "KEY", "None", "Used with window.FindElement and with return values to uniquely identify this element",),
+	("PSGVAL_CHECKBOX_METADATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "METADATA", "None", "User metadata that can be set to ANYTHING",),
+	("PSGVAL_CHECKBOX_PAD", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "PAD", "None", "Amount of padding to put around element (left/right, top/bottom) or ((left, right), (top, bottom))",),
+	("PSGVAL_CHECKBOX_S", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "S", "(None, None)", "(width, height) width = characters-wide, height = rows-high",),
+	("PSGVAL_CHECKBOX_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "SIZE", "(None, None)", "(width, height) width = characters-wide, height = rows-high",),
+	("PSGVAL_CHECKBOX_TEXT", FMAXPSG_SCTN0905_TUPDICTSTRADD, "CHECKBOX", "TEXT", "", "Text to display next to checkbox",),
+	("PSGVAL_CHECKBOX_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "TEXT_COLOR", "None", "color of the text",),
+	("PSGVAL_CHECKBOX_TOOLTIP", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "TOOLTIP", "None", "that will appear when mouse hovers over the element",),
+	("PSGVAL_CHECKBOX_VISIBLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "CHECKBOX", "VISIBLE", "True", "set visibility state of the element",),
+	("PSGVAL_COLOR_BACKGROUND", FMAXPSG_SCTN0900_STRDEF, "BACKGROUNDCOLOR", "#331122", "the background of the main frames",),
 	("PSGVAL_COLOR_BLACK", FMAXPSG_SCTN0900_STRDEF, "BLACK", "#000000", "black",),
-	("PSGVAL_COLOR_BLINK", FMAXPSG_SCTN0900_STRDEF, "BLINKCOLOR", "#CC0088", "the highlight color used in blinking bits when they are 'lit'",),
 	("PSGVAL_COLOR_GRAY3", FMAXPSG_SCTN0900_STRDEF, "GRAY3", "#333333", "gray 3",),
 	("PSGVAL_COLOR_GRAY6", FMAXPSG_SCTN0900_STRDEF, "GRAY6", "#666666", "gray 6",),
 	("PSGVAL_COLOR_GRAY9", FMAXPSG_SCTN0900_STRDEF, "GRAY9", "#999999", "gray 9",),
 	("PSGVAL_COLOR_GRAYC", FMAXPSG_SCTN0900_STRDEF, "GRAYC", "#CCCCCC", "gray C",),
-	("PSGVAL_COLOR_TIME", FMAXPSG_SCTN0900_STRDEF, "TIMECOLOR", "#660044", "the color the clock digits are",),
+	("PSGVAL_COLOR_NORMAL", FMAXPSG_SCTN0900_STRDEF, "NORMALCOLOR", "#660044", "the color the clock digits are",),
+	("PSGVAL_COLOR_NORMALHIGH", FMAXPSG_SCTN0900_STRDEF, "NORMALHIGH", "#CC0088", "the highlight color used in blinking bits when they are 'lit'",),
+	("PSGVAL_COLOR_NORMALLOW", FMAXPSG_SCTN0900_STRDEF, "NORMALLOW", "#330022", "the color the clock digits are",),
 	("PSGVAL_COLOR_WHITE", FMAXPSG_SCTN0900_STRDEF, "WHITE", "#FFFFFF", "white",),
-	("PSGVAL_WINDOW", FMAXPSG_SCTN0905_TUPDICTDEF, "WINDOW", "comment",),
-	("PSGVAL_WINDOW_ALPHA_CHANNEL", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ALPHA_CHANNEL", "1", "",),
-	("PSGVAL_WINDOW_AUTO_CLOSE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_CLOSE", "False", "",),
-	("PSGVAL_WINDOW_AUTO_CLOSE_DURATION", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_CLOSE_DURATION", "3", "",),
-	("PSGVAL_WINDOW_AUTO_SIZE_BUTTONS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_SIZE_BUTTONS", "None", "",),
-	("PSGVAL_WINDOW_AUTO_SIZE_TEXT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_SIZE_TEXT", "None", "",),
-	("PSGVAL_WINDOW_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "BACKGROUND_COLOR", "None", "",),
-	("PSGVAL_WINDOW_BORDER_DEPTH", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "BORDER_DEPTH", "None", "",),
-	("PSGVAL_WINDOW_BUTTON_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "BUTTON_COLOR", "None", "",),
-	("PSGVAL_WINDOW_DEBUGGER_ENABLED", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DEBUGGER_ENABLED", "True", "",),
-	("PSGVAL_WINDOW_DEFAULT_BUTTON_ELEMENT_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DEFAULT_BUTTON_ELEMENT_SIZE", "(None, None)", "",),
-	("PSGVAL_WINDOW_DEFAULT_ELEMENT_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DEFAULT_ELEMENT_SIZE", "(45, 1)", "",),
-	("PSGVAL_WINDOW_DISABLE_CLOSE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DISABLE_CLOSE", "False", "",),
-	("PSGVAL_WINDOW_DISABLE_MINIMIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DISABLE_MINIMIZE", "False", "",),
-	("PSGVAL_WINDOW_ELEMENT_JUSTIFICATION", FMAXPSG_SCTN0905_TUPDICTSTRADD, "WINDOW", "ELEMENT_JUSTIFICATION", "left", "",),
-	("PSGVAL_WINDOW_ELEMENT_PADDING", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ELEMENT_PADDING", "None", "",),
-	("PSGVAL_WINDOW_ENABLE_CLOSE_ATTEMPTED_EVENT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ENABLE_CLOSE_ATTEMPTED_EVENT", "False", "",),
-	("PSGVAL_WINDOW_FINALIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "FINALIZE", "False", "",),
-	("PSGVAL_WINDOW_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "FONT", "None", "",),
-	("PSGVAL_WINDOW_FORCE_TOPLEVEL", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "FORCE_TOPLEVEL", "False", "",),
-	("PSGVAL_WINDOW_GRAB_ANYWHERE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "GRAB_ANYWHERE", "False", "",),
-	("PSGVAL_WINDOW_ICON", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ICON", "None", "",),
-	("PSGVAL_WINDOW_KEEP_ON_TOP", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "KEEP_ON_TOP", "False", "",),
-	("PSGVAL_WINDOW_LAYOUT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "LAYOUT", "None", "",),
-	("PSGVAL_WINDOW_LOCATION", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "LOCATION", "(None, None)", "",),
-	("PSGVAL_WINDOW_MARGINS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "MARGINS", "(None, None)", "",),
-	("PSGVAL_WINDOW_METADATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "METADATA", "None", "",),
-	("PSGVAL_WINDOW_MODAL", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "MODAL", "False", "",),
-	("PSGVAL_WINDOW_NO_TITLEBAR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "NO_TITLEBAR", "False", "",),
-	("PSGVAL_WINDOW_PROGRESS_BAR_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "PROGRESS_BAR_COLOR", "(None, None)", "",),
-	("PSGVAL_WINDOW_RESIZABLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RESIZABLE", "False", "",),
-	("PSGVAL_WINDOW_RETURN_KEYBOARD_EVENTS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RETURN_KEYBOARD_EVENTS", "False", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU", "None", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_BACKGROUND_COLOR", "None", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_DISABLED_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_DISABLED_TEXT_COLOR", "None", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_FONT", "None", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_SELECTED_COLORS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_SELECTED_COLORS", "(None, None)", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_TEAROFF", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_TEAROFF", "False", "",),
-	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_TEXT_COLOR", "None", "",),
-	("PSGVAL_WINDOW_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "SIZE", "(None, None)", "",),
-	("PSGVAL_WINDOW_TEXT_JUSTIFICATION", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TEXT_JUSTIFICATION", "None", "",),
-	("PSGVAL_WINDOW_TITLE", FMAXPSG_SCTN0905_TUPDICTSTRADD, "WINDOW", "TITLE", "", "holds the window title",),
-	("PSGVAL_WINDOW_TITLEBAR_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_BACKGROUND_COLOR", "None", "",),
-	("PSGVAL_WINDOW_TITLEBAR_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_FONT", "None", "",),
-	("PSGVAL_WINDOW_TITLEBAR_ICON", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_ICON", "None", "",),
-	("PSGVAL_WINDOW_TITLEBAR_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_TEXT_COLOR", "None", "",),
-	("PSGVAL_WINDOW_TRANSPARENT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TRANSPARENT_COLOR", "None", "",),
-	("PSGVAL_WINDOW_TTK_THEME", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TTK_THEME", "None", "",),
-	("PSGVAL_WINDOW_USE_CUSTOM_TITLEBAR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "USE_CUSTOM_TITLEBAR", "None", "",),
-	("PSGVAL_WINDOW_USE_DEFAULT_FOCUS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "USE_DEFAULT_FOCUS", "True", "",),
-	("PSGVAL_WINDOW_USE_TTK_BUTTONS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "USE_TTK_BUTTONS", "None", "",),
+	("PSGVAL_WINDOW", FMAXPSG_SCTN0905_TUPDICTDEF, "WINDOW", "define the WINDOW tupdict",),
+	("PSGVAL_WINDOW_ALPHA_CHANNEL", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ALPHA_CHANNEL", "1", "Sets the opacity of the window. 0 = invisible 1 = completely visible. Values bewteen 0 & 1 will produce semi-transparent windows in SOME environments (The Raspberry Pi always has this value at 1 and cannot change.",),
+	("PSGVAL_WINDOW_AUTO_CLOSE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_CLOSE", "False", "If True, the window will automatically close itself",),
+	("PSGVAL_WINDOW_AUTO_CLOSE_DURATION", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_CLOSE_DURATION", "3", "Number of seconds to wait before closing the window",),
+	("PSGVAL_WINDOW_AUTO_SIZE_BUTTONS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_SIZE_BUTTONS", "None", "True if Buttons in this Window should be sized to exactly fit the text on this.",),
+	("PSGVAL_WINDOW_AUTO_SIZE_TEXT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "AUTO_SIZE_TEXT", "None", "True if Elements in Window should be sized to exactly fir the length of text",),
+	("PSGVAL_WINDOW_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "BACKGROUND_COLOR", "None", "color of background",),
+	("PSGVAL_WINDOW_BORDER_DEPTH", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "BORDER_DEPTH", "None", "Default border depth (width) for all elements in the window",),
+	("PSGVAL_WINDOW_BUTTON_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "BUTTON_COLOR", "None", "Default button colors for all buttons in the window",),
+	("PSGVAL_WINDOW_DEBUGGER_ENABLED", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DEBUGGER_ENABLED", "True", "Default border depth (width) for all elements in the window",),
+	("PSGVAL_WINDOW_DEFAULT_BUTTON_ELEMENT_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DEFAULT_BUTTON_ELEMENT_SIZE", "(None, None)", "(width, height) size in characters (wide) and rows (high) for all Button elements in this window",),
+	("PSGVAL_WINDOW_DEFAULT_ELEMENT_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DEFAULT_ELEMENT_SIZE", "(45, 1)", "size in characters (wide) and rows (high) for all elements in this window",),
+	("PSGVAL_WINDOW_DISABLE_CLOSE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DISABLE_CLOSE", "False", "If True, the X button in the top right corner of the window will no work.  Use with caution and always give a way out toyour users",),
+	("PSGVAL_WINDOW_DISABLE_MINIMIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "DISABLE_MINIMIZE", "False", "if True the user won't be able to minimize window.  Good for taking over entire screen and staying that way.",),
+	("PSGVAL_WINDOW_ELEMENT_JUSTIFICATION", FMAXPSG_SCTN0905_TUPDICTSTRADD, "WINDOW", "ELEMENT_JUSTIFICATION", "left", "All elements in the Window itself will have this justification 'left', 'right', 'center' are valid values",),
+	("PSGVAL_WINDOW_ELEMENT_PADDING", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ELEMENT_PADDING", "None", "Default amount of padding to put around elements in window (left/right, top/bottom) or ((left, right), (top, bottom))",),
+	("PSGVAL_WINDOW_ENABLE_CLOSE_ATTEMPTED_EVENT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ENABLE_CLOSE_ATTEMPTED_EVENT", "False", "If True then the window will not close when 'X' clicked. Instead an event WINDOW_CLOSE_ATTEMPTED_EVENT if returned from window.read",),
+	("PSGVAL_WINDOW_FINALIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "FINALIZE", "False", "If True then the Finalize method will be called. Use this rather than chaining .Finalize for cleaner code",),
+	("PSGVAL_WINDOW_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "FONT", "None", "specifies the font family, size, etc",),
+	("PSGVAL_WINDOW_FORCE_TOPLEVEL", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "FORCE_TOPLEVEL", "False", "If True will cause this window to skip the normal use of a hidden master window",),
+	("PSGVAL_WINDOW_GRAB_ANYWHERE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "GRAB_ANYWHERE", "False", "If True can use mouse to click and drag to move the window. Almost every location of the window will work except input fields on some systems",),
+	("PSGVAL_WINDOW_ICON", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "ICON", "None", "Can be either a filename or Base64 value. For Windows if filename, it MUST be ICO format. For Linux, must NOT be ICO",),
+	("PSGVAL_WINDOW_KEEP_ON_TOP", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "KEEP_ON_TOP", "False", "If True, window will be created on top of all other windows on screen. It can be bumped down if another window created with this parm",),
+	("PSGVAL_WINDOW_LAYOUT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "LAYOUT", "None", "The layout for the window. Can also be specified in the Layout method",),
+	("PSGVAL_WINDOW_LOCATION", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "LOCATION", "(None, None)", "(x,y) location, in pixels, to locate the upper left corner of the window on the screen. Default is to center on screen.",),
+	("PSGVAL_WINDOW_MARGINS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "MARGINS", "(None, None)", "(left/right, top/bottom) Amount of pixels to leave inside the window's frame around the edges before your elements are shown.",),
+	("PSGVAL_WINDOW_METADATA", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "METADATA", "None", "User metadata that can be set to ANYTHING",),
+	("PSGVAL_WINDOW_MODAL", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "MODAL", "False", "If True then this window will be the only window a user can interact with until it is closed",),
+	("PSGVAL_WINDOW_NO_TITLEBAR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "NO_TITLEBAR", "False", "If true, no titlebar nor frame will be shown on window. This means you cannot minimize the window and it will not show up on the taskbar",),
+	("PSGVAL_WINDOW_PROGRESS_BAR_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "PROGRESS_BAR_COLOR", "(None, None)", "(bar color, background color) Sets the default colors for all progress bars in the window",),
+	("PSGVAL_WINDOW_RESIZABLE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RESIZABLE", "False", "If True, allows the user to resize the window. Note the not all Elements will change size or location when resizing.",),
+	("PSGVAL_WINDOW_RETURN_KEYBOARD_EVENTS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RETURN_KEYBOARD_EVENTS", "False", "if True key presses on the keyboard will be returned as Events from Read calls",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU", "None", "A list of lists of Menu items to show when this element is right clicked. See user docs for exact format.",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_BACKGROUND_COLOR", "None", "Background color for right click menus",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_DISABLED_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_DISABLED_TEXT_COLOR", "None", "Text color for disabled right click menu items",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_FONT", "None", "Font for right click menus",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_SELECTED_COLORS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_SELECTED_COLORS", "(None, None)", "Text AND background colors for a selected item. Can be a Tuple OR a color string. simplified-button-color-string 'foreground on background'. Can be a single color if want to set only the background.",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_TEAROFF", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_TEAROFF", "False", "If True then all right click menus can be torn off",),
+	("PSGVAL_WINDOW_RIGHT_CLICK_MENU_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "RIGHT_CLICK_MENU_TEXT_COLOR", "None", "Text color for right click menus",),
+	("PSGVAL_WINDOW_SIZE", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "SIZE", "(None, None)", "(width, height) size in pixels for this window. Normally the window is autosized to fit contents, not set to an absolute size by the user",),
+	("PSGVAL_WINDOW_TEXT_JUSTIFICATION", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TEXT_JUSTIFICATION", "None", "Default text justification for all Text Elements in window",),
+	("PSGVAL_WINDOW_TITLE", FMAXPSG_SCTN0905_TUPDICTSTRADD, "WINDOW", "TITLE", "", "The title that will be displayed in the Titlebar and on the Taskbar",),
+	("PSGVAL_WINDOW_TITLEBAR_BACKGROUND_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_BACKGROUND_COLOR", "None", "If custom titlebar indicated by use_custom_titlebar, then use this as background color",),
+	("PSGVAL_WINDOW_TITLEBAR_FONT", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_FONT", "None", "If custom titlebar indicated by use_custom_titlebar, then use this as title font",),
+	("PSGVAL_WINDOW_TITLEBAR_ICON", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_ICON", "None", "If custom titlebar indicated by use_custom_titlebar, then use this as the icon (file or base64 bytes)",),
+	("PSGVAL_WINDOW_TITLEBAR_TEXT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TITLEBAR_TEXT_COLOR", "None", "If custom titlebar indicated by use_custom_titlebar, then use this as text color",),
+	("PSGVAL_WINDOW_TRANSPARENT_COLOR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TRANSPARENT_COLOR", "None", "Any portion of the window that has this color will be completely transparent. You can even click through these spots to the window under this window.",),
+	("PSGVAL_WINDOW_TTK_THEME", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "TTK_THEME", "None", "Set the tkinter ttk 'theme' of the window.  Default = DEFAULT_TTK_THEME.  Sets all ttk widgets to this theme as their default",),
+	("PSGVAL_WINDOW_USE_CUSTOM_TITLEBAR", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "USE_CUSTOM_TITLEBAR", "None", "If True, then a custom titlebar will be used instead of the normal titlebar",),
+	("PSGVAL_WINDOW_USE_DEFAULT_FOCUS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "USE_DEFAULT_FOCUS", "True", "If True will use the default focus algorithm to set the focus to the 'Correct' element",),
+	("PSGVAL_WINDOW_USE_TTK_BUTTONS", FMAXPSG_SCTN0905_TUPDICTVALADD, "WINDOW", "USE_TTK_BUTTONS", "None", "Affects all buttons in window. True = use ttk buttons. False = do not use ttk buttons.  None = use ttk buttons only if on a Mac",),
 	("PSGVAL__ALPHA_CHANNEL", FMAXPSG_SCTN0901_STRDEF, "ALPHA_CHANNEL", "alpha_channel", "",),
 	("PSGVAL__AUTO_CLOSE", FMAXPSG_SCTN0901_STRDEF, "AUTO_CLOSE", "auto_close", "",),
 	("PSGVAL__AUTO_CLOSE_DURATION", FMAXPSG_SCTN0901_STRDEF, "AUTO_CLOSE_DURATION", "auto_close_duration", "",),
 	("PSGVAL__AUTO_SIZE_BUTTON", FMAXPSG_SCTN0901_STRDEF, "AUTO_SIZE_BUTTON", "auto_size_button", "",),
 	("PSGVAL__AUTO_SIZE_BUTTONS", FMAXPSG_SCTN0901_STRDEF, "AUTO_SIZE_BUTTONS", "auto_size_buttons", "",),
+	("PSGVAL__AUTO_SIZE_TEXT", FMAXPSG_SCTN0901_STRDEF, "AUTO_SIZE_TEXT", "auto_size_text", "",),
 	("PSGVAL__AUTO_SIZE_TEXT", FMAXPSG_SCTN0901_STRDEF, "AUTO_SIZE_TEXT", "auto_size_text", "",),
 	("PSGVAL__BACKGROUND_COLOR", FMAXPSG_SCTN0901_STRDEF, "BACKGROUND_COLOR", "background_color", "",),
 	("PSGVAL__BIND_RETURN_KEY", FMAXPSG_SCTN0901_STRDEF, "BIND_RETURN_KEY", "bind_return_key", "",),
@@ -601,7 +632,9 @@ TBGLST = [
 	("PSGVAL__BUTTON_TEXT", FMAXPSG_SCTN0901_STRDEF, "BUTTON_TEXT", "button_text", "",),
 	("PSGVAL__BUTTON_TYPE", FMAXPSG_SCTN0901_STRDEF, "BUTTON_TYPE", "button_type", "",),
 	("PSGVAL__CHANGE_SUBMITS", FMAXPSG_SCTN0901_STRDEF, "CHANGE_SUBMITS", "change_submits", "",),
+	("PSGVAL__CHECKBOX_COLOR", FMAXPSG_SCTN0901_STRDEF, "CHECKBOX_COLOR", "checkbox_color", "color of background of the box that has the check mark in it. The checkmark is the same color as the text",),
 	("PSGVAL__DEBUGGER_ENABLED", FMAXPSG_SCTN0901_STRDEF, "DEBUGGER_ENABLED", "debugger_enabled", "",),
+	("PSGVAL__DEFAULT", FMAXPSG_SCTN0901_STRDEF, "DEFAULT", "default", "",),
 	("PSGVAL__DEFAULT_BUTTON_ELEMENT_SIZE", FMAXPSG_SCTN0901_STRDEF, "DEFAULT_BUTTON_ELEMENT_SIZE", "default_button_element_size", "",),
 	("PSGVAL__DEFAULT_ELEMENT_SIZE", FMAXPSG_SCTN0901_STRDEF, "DEFAULT_ELEMENT_SIZE", "default_element_size", "",),
 	("PSGVAL__DEFAULT_EXTENSION", FMAXPSG_SCTN0901_STRDEF, "DEFAULT_EXTENSION", "default_extension", "",),
@@ -649,6 +682,8 @@ TBGLST = [
 	("PSGVAL__S", FMAXPSG_SCTN0901_STRDEF, "S", "s", "",),
 	("PSGVAL__SIZE", FMAXPSG_SCTN0901_STRDEF, "SIZE", "size", "",),
 	("PSGVAL__TARGET", FMAXPSG_SCTN0901_STRDEF, "TARGET", "target", "",),
+	("PSGVAL__TEXT", FMAXPSG_SCTN0901_STRDEF, "TEXT", "text", "",),
+	("PSGVAL__TEXT_COLOR", FMAXPSG_SCTN0901_STRDEF, "TEXT_COLOR", "text_color", "",),
 	("PSGVAL__TEXT_JUSTIFICATION", FMAXPSG_SCTN0901_STRDEF, "TEXT_JUSTIFICATION", "text_justification", "",),
 	("PSGVAL__TITLE", FMAXPSG_SCTN0901_STRDEF, "TITLE", "title", "",),
 	("PSGVAL__TITLEBAR_BACKGROUND_COLOR", FMAXPSG_SCTN0901_STRDEF, "TITLEBAR_BACKGROUND_COLOR", "titlebar_background_color", "",),
@@ -662,6 +697,9 @@ TBGLST = [
 	("PSGVAL__USE_DEFAULT_FOCUS", FMAXPSG_SCTN0901_STRDEF, "USE_DEFAULT_FOCUS", "use_default_focus", "",),
 	("PSGVAL__USE_TTK_BUTTONS", FMAXPSG_SCTN0901_STRDEF, "USE_TTK_BUTTONS", "use_ttk_buttons", "",),
 	("PSGVAL__VISIBLE", FMAXPSG_SCTN0901_STRDEF, "VISIBLE", "visible", "",),
+	("PSGVAL__CLOCKTIME", FMAXPSG_SCTN0901_KEYDEF, "CLOCKTIME", "holds the clock value (str HH:MM:SS)",),
+	("PSGVAL__ELAPSEDTIME", FMAXPSG_SCTN0901_KEYDEF, "ELAPSEDTIME", "holds the elapsed time",),
+	("PSGVAL__NEXTTIME", FMAXPSG_SCTN0901_KEYDEF, "NEXTTIME", "holds the next scheduled time countdown",),
 	("PSGVAL_____", FMAX_NOP, "FMPSG_BEGINS",),
 	# fold here ⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1⥣1
 ]
@@ -1008,7 +1046,7 @@ def makePSG():
 	strToRtn_ = ""
 
 	# ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣ ⥥1⥣
-	strToRtn_ += f"""{makeAComment("SCTN0900 DEF1")}"""
+	strToRtn_ += f"""{readFileToStr(f"{PSGTOP_NAME}")}{makeAComment("SCTN0900 DEF1")}"""
 	dictToUse_ = sortADict(FMPSG_SCTN0900_DEF1DICT)
 	for thisName_, value_ in dictToUse_.items():
 		strToRtn_ += f"""{thisName_} = {value_}  # {FMPSG_SCTN0900_DEF1CMNTDICT[thisName_]}{NEWLINE}"""
@@ -1048,6 +1086,8 @@ def makePSG():
 	dictToUse_ = sortADict(FMPSG_SCTN0905_TUPDICTDICT)
 	for thisName_, value_ in dictToUse_.items():
 		strToRtn_ += f"""{makeANormalTupDict(thisName_, value_)}"""
+
+	strToRtn_ += f"""{makeAWideComment("end of managed sections of PSG.py")}"""
 	strToRtn_ += f"""{NEWLINE}{NEWLINE}"""
 
 	return strToRtn_
