+++ PSG.py
@@ -6,1611 +6,332 @@ import PySimpleGUI as SG
 import CF
 
 
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# setting constants
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-
-BLACK = "#000000"
-GRAY8 = "#888888"
-GRAYC = "#CCCCCC"
-GREEN2 = "#448811"
-MAINDOWNCOLOR = "#880000"
-MAINDOWNTEXTCOLOR = "#FF0000"
-MAINUPCOLOR = "#008800"
-MAINUPTEXTCOLOR = "#00FF00"
-ORANGE = "#FF7F00"
-PURP440022 = "#440022"
-PURP660044 = "#660044"
-TEAL1 = "#00FF7F"
-YELLOW2 = "#444422"
-YELLOW666600 = "#666600"
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * SCTN0900 DEF1
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+BACKGROUNDCOLOR = "#331122"  # the background of the main frames
+BLACK = "#000000"  # black
+GRAY3 = "#333333"  # gray 3
+GRAY6 = "#666666"  # gray 6
+GRAY9 = "#999999"  # gray 9
+GRAYC = "#CCCCCC"  # gray C
+NORMALCOLOR = "#660044"  # the color the clock digits are
+NORMALHIGH = "#CC0088"  # the highlight color used in blinking bits when they are 'lit'
+NORMALLOW = "#330022"  # the color the clock digits are
+WHITE = "#FFFFFF"  # white
 
 
-AIR = "AIR"
-DYNAVAP = "DYNAVAP"
-FIREFLY = "FIREFLY"
-PEAK = "PEAK"
-Q = "Q"
-QDDM = "QDDM"
-QOMO = "QOMO"
-QTZBANGER = "QTZBANGER"
-VARIOUS = "VARIOUS"
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * SCTN0901 DEF2
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+ALPHA_CHANNEL = "alpha_channel"  #
+AUTO_CLOSE = "auto_close"  #
+AUTO_CLOSE_DURATION = "auto_close_duration"  #
+AUTO_SIZE_BUTTON = "auto_size_button"  #
+AUTO_SIZE_BUTTONS = "auto_size_buttons"  #
+AUTO_SIZE_TEXT = "auto_size_text"  #
+BACKGROUND_COLOR = "background_color"  #
+BIND_RETURN_KEY = "bind_return_key"  #
+BORDER_DEPTH = "border_depth"  #
+BORDER_WIDTH = "border_width"  #
+BUTTON_COLOR = "button_color"  #
+BUTTON_TEXT = "button_text"  #
+BUTTON_TYPE = "button_type"  #
+CHANGE_SUBMITS = "change_submits"  #
+CHECKBOX_COLOR = "checkbox_color"  # color of background of the box that has the check mark in it. The checkmark is the same color as the text
+CLOCKTIME = "CLOCKTIME"  # holds the clock value (str HH:MM:SS)
+DEBUGGER_ENABLED = "debugger_enabled"  #
+DEFAULT = "default"  #
+DEFAULT_BUTTON_ELEMENT_SIZE = "default_button_element_size"  #
+DEFAULT_ELEMENT_SIZE = "default_element_size"  #
+DEFAULT_EXTENSION = "default_extension"  #
+DISABLE_CLOSE = "disable_close"  #
+DISABLE_MINIMIZE = "disable_minimize"  #
+DISABLED = "disabled"  #
+DISABLED_BUTTON_COLOR = "disabled_button_color"  #
+ELAPSEDTIME = "ELAPSEDTIME"  # holds the elapsed time
+ELEMENT_JUSTIFICATION = "element_justification"  #
+ELEMENT_PADDING = "element_padding"  #
+ENABLE_CLOSE_ATTEMPTED_EVENT = "enable_close_attempted_event"  #
+ENABLE_EVENTS = "enable_events"  #
+FILE_TYPES = "file_types"  #
+FINALIZE = "finalize"  #
+FOCUS = "focus"  #
+FONT = "font"  #
+FORCE_TOPLEVEL = "force_toplevel"  #
+GRAB_ANYWHERE = "grab_anywhere"  #
+HIGHLIGHT_COLORS = "highlight_colors"  #
+ICON = "icon"  #
+IMAGE_DATA = "image_data"  #
+IMAGE_FILENAME = "image_filename"  #
+IMAGE_SIZE = "image_size"  #
+IMAGE_SUBSAMPLE = "image_subsample"  #
+INITIAL_FOLDER = "initial_folder"  #
+K = "k"  #
+KEEP_ON_TOP = "keep_on_top"  #
+KEY = "key"  #
+LAYOUT = "layout"  #
+LOCATION = "location"  #
+MARGINS = "margins"  #
+METADATA = "metadata"  #
+MODAL = "modal"  #
+NEXTTIME = "NEXTTIME"  # holds the next scheduled time countdown
+NO_TITLEBAR = "no_titlebar"  #
+PAD = "pad"  #
+PROGRESS_BAR_COLOR = "progress_bar_color"  #
+RESIZABLE = "resizable"  #
+RETURN_KEYBOARD_EVENTS = "return_keyboard_events"  #
+RIGHT_CLICK_MENU = "right_click_menu"  #
+RIGHT_CLICK_MENU_BACKGROUND_COLOR = "right_click_menu_background_color"  #
+RIGHT_CLICK_MENU_DISABLED_TEXT_COLOR = "right_click_menu_disabled_text_color"  #
+RIGHT_CLICK_MENU_FONT = "right_click_menu_font"  #
+RIGHT_CLICK_MENU_SELECTED_COLORS = "right_click_menu_selected_colors"  #
+RIGHT_CLICK_MENU_TEAROFF = "right_click_menu_tearoff"  #
+RIGHT_CLICK_MENU_TEXT_COLOR = "right_click_menu_text_color"  #
+S = "s"  #
+SIZE = "size"  #
+TARGET = "target"  #
+TEXT = "text"  #
+TEXT_COLOR = "text_color"  #
+TEXT_JUSTIFICATION = "text_justification"  #
+TITLE = "title"  #
+TITLEBAR_BACKGROUND_COLOR = "titlebar_background_color"  #
+TITLEBAR_FONT = "titlebar_font"  #
+TITLEBAR_ICON = "titlebar_icon"  #
+TITLEBAR_TEXT_COLOR = "titlebar_text_color"  #
+TOOLTIP = "tooltip"  #
+TRANSPARENT_COLOR = "transparent_color"  #
+TTK_THEME = "ttk_theme"  #
+USE_CUSTOM_TITLEBAR = "use_custom_titlebar"  #
+USE_DEFAULT_FOCUS = "use_default_focus"  #
+USE_TTK_BUTTONS = "use_ttk_buttons"  #
+VISIBLE = "visible"  #
 
 
-TASKLIST = [
-	AIR,
-	DYNAVAP,
-	FIREFLY,
-	PEAK,
-	Q,
-	QDDM,
-	QOMO,
-	QTZBANGER,
-	VARIOUS,
-]
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * SCTN0902 dicts
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 
 
-ADJBTNDOWNCOLOR = MAINDOWNCOLOR
-ADJBTNDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
-ADJBTNUPCOLOR = MAINUPCOLOR
-ADJBTNUPTEXTCOLOR = MAINUPTEXTCOLOR
-ADJTIMEDOWNBKGNDCOLOR = MAINDOWNCOLOR
-ADJTIMETXTCOLOR = GRAYC
-ADJTIMEUPBKGNDCOLOR = MAINUPCOLOR
-BTNDOWNCOLOR = MAINDOWNCOLOR
-BTNDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
-BTNQUITCOLOR = PURP440022
-BTNQUITTEXTCOLOR = YELLOW666600
-BTNRESETONECOLOR = MAINDOWNCOLOR
-BTNRESETONETEXTCOLOR = MAINDOWNTEXTCOLOR
-BTNTASKDOWNCOLOR = MAINDOWNCOLOR
-BTNTASKDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
-BTNTASKUPCOLOR = MAINUPCOLOR
-BTNTASKUPTEXTCOLOR = MAINUPTEXTCOLOR
-BTNUPCOLOR = MAINUPCOLOR
-BTNUPNOTIMECOLOR = MAINUPCOLOR
-BTNUPNOTIMETEXTCOLOR = MAINUPTEXTCOLOR
-BTNUPTEXTCOLOR = MAINUPTEXTCOLOR
-BTNZEROCOLOR = PURP440022
-BTNZEROTEXTCOLOR = YELLOW666600
-INTERRUPTRATE = 100
-LASTFILENAME = "biditi.last"
-MAXTICKS = 59999
-MODE_NORMAL = "MODE_NORMAL"
-MODE_RESTART = "MODE_RESTART"
-MODE_START = "MODE_START"
-NOT = "NOT"
-SPACECOLOR = GRAY8
-SPINBKGNDCOLOR = TEAL1
-SPINSIZE = (10, 1)
-SPINTEXTCOLOR = ORANGE
-TASKCOUNTERCOLOR = GREEN2
-TASKCOUNTSZ = (3, 1)
-TICKSPERSEC = 10
-TIMERDOWNBKGNDCOLOR = MAINDOWNCOLOR
-TIMERDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
-TIMEROFFBKGNDCOLOR = BLACK
-TIMEROFFTXTCOLOR = YELLOW2
-TIMERUPBKGNDCOLOR = MAINUPCOLOR
-TIMERUPTEXTCOLOR = MAINUPTEXTCOLOR
-PCADJTIMEBTNFONTSZ = 6
-PCADJTIMEDOWNIMAGE = DN32
-PCADJTIMEFONTSZ = 18
-PCADJTIMEUPIMAGE = UP32
-PCBOWLSDOWN = BDOWN32
-PCBOWLSFONTSZ = 18
-PCBOWLSUP = BUP32
-PCBTNFONTSZ = 10
-PCCOUNTERFONTSZ = 30
-PCDOWNIMAGE = DN64
-PCFONT = "Source Code Pro"
-PCLABELFONTSZ = 12
-PCQUITBTN = QUIT64
-PCRESETONE = RESETONE64
-PCSETTIMERFONTSZ = 20
-PCSPACEFONTSZ = 10
-PCSPINFONTSZ = 16
-PCTIMERFONTSZ = 130
-PCUPDOWNTIMEFONTSZ = 50
-PCUPIMAGE = UP64
-PCUPNOTIME = UPNOTIME64
-PCZEROALL = ZEROALL64
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * SCTN0903 lists
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 
 
-CWD = PATH.abspath(".")
-if CWD.find("_android") > -1:
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	ADJTIMEBTNFONTSZ = 6
-	ADJTIMEDOWNIMAGE = DN64
-	ADJTIMEFONTSZ = 10
-	ADJTIMEUPIMAGE = UP64
-	BOWLSDOWN = BDOWN64
-	BOWLSFONTSZ = 10
-	BOWLSUP = BUP64
-	BTNFONTSZ = 12
-	CONFIGDIRECTORY = ""
-	COUNTERFONTSZ = 20
-	DOWNIMAGE = DN128
-	FONT = "Source Code Pro"
-	LABELFONTSZ = 12
-	QUITBTN = QUIT128
-	RESETONE = RESETONE128
-	SETTIMERFONTSZ = 20
-	SG.ChangeLookAndFeel("DarkPurple6")
-	SPACEFONTSZ = 10
-	SPINFONTSZ = 9
-	TIMERFONTSZ = 70
-	UPDOWNTIMEFONTSZ = 34
-	UPIMAGE = UP128
-	UPNOTIME = UPNOTIME128
-	ZEROALL = ZEROALL128
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-elif CWD.find("_DEV") > -1:
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	ADJTIMEBTNFONTSZ = PCADJTIMEBTNFONTSZ
-	ADJTIMEDOWNIMAGE = PCADJTIMEDOWNIMAGE
-	ADJTIMEFONTSZ = PCADJTIMEFONTSZ
-	ADJTIMEUPIMAGE = PCADJTIMEUPIMAGE
-	BOWLSDOWN = PCBOWLSDOWN
-	BOWLSFONTSZ = PCBOWLSFONTSZ
-	BOWLSUP = PCBOWLSUP
-	BTNFONTSZ = PCBTNFONTSZ
-	CONFIGDIRECTORY = "/home/will/.config/biditi_DEV/"
-	COUNTERFONTSZ = PCCOUNTERFONTSZ
-	DOWNIMAGE = PCDOWNIMAGE
-	FONT = PCFONT
-	LABELFONTSZ = PCLABELFONTSZ
-	QUITBTN = PCQUITBTN
-	RESETONE = PCRESETONE
-	SETTIMERFONTSZ = PCSETTIMERFONTSZ
-	SG.ChangeLookAndFeel("DarkGreen5")
-	SPACEFONTSZ = PCSPACEFONTSZ
-	SPINFONTSZ = PCSPINFONTSZ
-	TIMERFONTSZ = PCTIMERFONTSZ
-	UPDOWNTIMEFONTSZ = PCUPDOWNTIMEFONTSZ
-	UPIMAGE = PCUPIMAGE
-	UPNOTIME = PCUPNOTIME
-	ZEROALL = PCZEROALL
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-else:
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	ADJTIMEBTNFONTSZ = PCADJTIMEBTNFONTSZ
-	ADJTIMEDOWNIMAGE = PCADJTIMEDOWNIMAGE
-	ADJTIMEFONTSZ = PCADJTIMEFONTSZ
-	ADJTIMEUPIMAGE = PCADJTIMEUPIMAGE
-	BOWLSDOWN = PCBOWLSDOWN
-	BOWLSFONTSZ = PCBOWLSFONTSZ
-	BOWLSUP = PCBOWLSUP
-	BTNFONTSZ = PCBTNFONTSZ
-	CONFIGDIRECTORY = "/home/will/.config/biditi/"
-	COUNTERFONTSZ = PCCOUNTERFONTSZ
-	DOWNIMAGE = PCDOWNIMAGE
-	FONT = PCFONT
-	LABELFONTSZ = PCLABELFONTSZ
-	QUITBTN = PCQUITBTN
-	RESETONE = PCRESETONE
-	SETTIMERFONTSZ = PCSETTIMERFONTSZ
-	SG.ChangeLookAndFeel("DarkPurple6")
-	SPACEFONTSZ = PCSPACEFONTSZ
-	SPINFONTSZ = PCSPINFONTSZ
-	TIMERFONTSZ = PCTIMERFONTSZ
-	UPDOWNTIMEFONTSZ = PCUPDOWNTIMEFONTSZ
-	UPIMAGE = PCUPIMAGE
-	UPNOTIME = PCUPNOTIME
-	ZEROALL = PCZEROALL
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * SCTN0904 platform equalizers
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 
 
-if CWD.find("_DEV") > -1:
-	DEVMODE = True
-	MAINWINDOWTITLE = "BIDITI_DEV"
-else:
-	DEVMODE = False
-	MAINWINDOWTITLE = "biditi"
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * SCTN0905 tupdict
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * start of BUTTON structures
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+
+BUTTONTUP = (
+	(AUTO_SIZE_BUTTON, None),  # if True the button size is sized to fit the text
+	(BIND_RETURN_KEY, False),  # If True the return key will cause this button to be pressed
+	(BORDER_WIDTH, None),  # width of border around button in pixels
+	(BUTTON_COLOR, None),  # Color of button. default is from theme or the window. Easy to remember which is which if you say 'ON' between colors. 'red' on 'green'. Normally a tuple, but can be a simplified-button-color-string 'foreground on background'. Can be a single color if want to set only the background.
+	(BUTTON_TEXT, ""),  # str text to display on the button
+	(BUTTON_TYPE, 7),  # You  should NOT be setting this directly. ONLY the shortcut functions set this
+	(CHANGE_SUBMITS, False),  # DO NOT USE. Only listed for backwards compat - Use enable_events instead
+	(DEFAULT_EXTENSION, ""),  # If no extension entered by user, add this to filename (only used in saveas dialogs)
+	(DISABLED, False),  # If True button will be created disabled. If BUTTON_DISABLED_MEANS_IGNORE then the button will be ignored rather than disabled using tkinter
+	(DISABLED_BUTTON_COLOR, None),  # colors to use when button is disabled (text, background). Use None for a color if don't want to change. Only ttk buttons support both text and background colors. tk buttons only support changing text color
+	(ENABLE_EVENTS, False),  # Turns on the element specific events. If this button is a target, should it generate an event when filled in
+	(FILE_TYPES, (('ALL FILES', '*.*'),)),  # the filetypes that will be used to match files. To indicate all files: (('ALL Files', '*.*'),).  Note - NOT SUPPORTED ON MAC
+	(FOCUS, False),  # if True, initial focus will be put on this button
+	(FONT, None),  # specifies the font family, size, etc
+	(HIGHLIGHT_COLORS, None),  # colors to use when button has focus (highlight, background). None will use computed colors. Only used by Linux and only for non-TTK button
+	(IMAGE_DATA, None),  # Raw or Base64 representation of the image to put on button. Choose either filename or data
+	(IMAGE_FILENAME, None),  # image filename if there is a button image. GIFs and PNGs only.
+	(IMAGE_SIZE, (None, None)),  # Size of the image in pixels (width, height)
+	(IMAGE_SUBSAMPLE, None),  # amount to reduce the size of the image. Divides the size by this number. 2=1/2, 3=1/3, 4=1/4, etc
+	(INITIAL_FOLDER, None),  # starting path for folders and files
+	(K, None),  # Used with window.FindElement and with return values to uniquely identify this element to uniquely identify this element
+	(KEY, None),  # Used with window.FindElement and with return values to uniquely identify this element to uniquely identify this element
+	(METADATA, None),  # User metadata that can be set to ANYTHING
+	(PAD, None),  # Amount of padding to put around element (left/right, top/bottom) or ((left, right), (top, bottom))
+	(RIGHT_CLICK_MENU, None),  # A list of lists of Menu items to show when this element is right clicked. See user docs for exact format.
+	(S, (None, None)),  # (width, height) of the button in characters wide, rows high
+	(SIZE, (None, None)),  # (width, height) of the button in characters wide, rows high
+	(TARGET, (None, None)),  # str | Tuple[int, int] key or (row,col) target for the button. Note that -1 for column means 1 element to the left of this one. The constant ThisRow is used to indicate the current row. The Button itself is a valid target for some types of button
+	(TOOLTIP, None),  # text, that will appear when mouse hovers over the element
+	(USE_TTK_BUTTONS, None),  # True = use ttk buttons. False = do not use ttk buttons.  None (Default) = use ttk buttons only if on a Mac and not with button images
+	(VISIBLE, True),  # set visibility state of the element
+)
+
+def BUTTONDICT():
+	return dict((x, y) for x, y in BUTTONTUP)
 
 
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# currentData dict keys and uses
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * start of CHECKBOX structures
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
 
-BUTTON = "BUTTON"
-DOWNSEC = "DOWNSEC"
-EVENTS = "EVENTS"
-FILENAME = "FILENAME"
-TASK1BOWLSCOUNT = "TASK1BOWLSCOUNT"
-TASK1COUNT = "TASK1COUNT"
-TASK1DOWNTIMER = "TASK1DOWNTIMER"
-TASK1NAME = "TASK1NAME"
-TASK1UPTIMER = "TASK1UPTIMER"
-TASK2BOWLSCOUNT = "TASK2BOWLSCOUNT"
-TASK2COUNT = "TASK2COUNT"
-TASK2DOWNTIMER = "TASK2DOWNTIMER"
-TASK2NAME = "TASK2NAME"
-TASK2UPTIMER = "TASK2UPTIMER"
-TASK3BOWLSCOUNT = "TASK3BOWLSCOUNT"
-TASK3COUNT = "TASK3COUNT"
-TASK3DOWNTIMER = "TASK3DOWNTIMER"
-TASK3NAME = "TASK3NAME"
-TASK3UPTIMER = "TASK3UPTIMER"
-TASK4BOWLSCOUNT = "TASK4BOWLSCOUNT"
-TASK4COUNT = "TASK4COUNT"
-TASK4DOWNTIMER = "TASK4DOWNTIMER"
-TASK4NAME = "TASK4NAME"
-TASK4UPTIMER = "TASK4UPTIMER"
-TEXTNAME = "TEXTNAME"
-UPSEC = "UPSEC"
+CHECKBOXTUP = (
+	(AUTO_SIZE_TEXT, None),  # if True will size the element to match the length of the text
+	(BACKGROUND_COLOR, None),  # color of background
+	(CHANGE_SUBMITS, False),  # DO NOT USE. Only listed for backwards compat - Use enable_events instead
+	(CHECKBOX_COLOR, None),  # color of background of the box that has the check mark in it. The checkmark is the same color as the text
+	(DEFAULT, False),  # Set to True if you want this checkbox initially checked
+	(DISABLED, False),  # set disable state
+	(ENABLE_EVENTS, False),  # Turns on the element specific events. Checkbox events happen when an item changes
+	(FONT, None),  # specifies the font family, size, etc
+	(K, None),  # Used with window.FindElement and with return values to uniquely identify this element
+	(KEY, None),  # Used with window.FindElement and with return values to uniquely identify this element
+	(METADATA, None),  # User metadata that can be set to ANYTHING
+	(PAD, None),  # Amount of padding to put around element (left/right, top/bottom) or ((left, right), (top, bottom))
+	(S, (None, None)),  # (width, height) width = characters-wide, height = rows-high
+	(SIZE, (None, None)),  # (width, height) width = characters-wide, height = rows-high
+	(TEXT, ""),  # Text to display next to checkbox
+	(TEXT_COLOR, None),  # color of the text
+	(TOOLTIP, None),  # that will appear when mouse hovers over the element
+	(VISIBLE, True),  # set visibility state of the element
+)
+
+def CHECKBOXDICT():
+	return dict((x, y) for x, y in CHECKBOXTUP)
 
 
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# the defaults for all things settable via currentData
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-DEFAULTS = [
-	(DOWNSEC, 7,),
-	(EVENTS, [],),
-	(FILENAME, "biditi.pkl",),
-	(TASK1BOWLSCOUNT, 0,),
-	(TASK1COUNT, 0,),
-	(TASK1DOWNTIMER, 7,),
-	(TASK1NAME, Q,),
-	(TASK1UPTIMER, 20,),
-	(TASK2BOWLSCOUNT, 0,),
-	(TASK2COUNT, 0,),
-	(TASK2DOWNTIMER, 7,),
-	(TASK2NAME, QTZBANGER,),
-	(TASK2UPTIMER, 20,),
-	(TASK3BOWLSCOUNT, 0,),
-	(TASK3COUNT, 0,),
-	(TASK3DOWNTIMER, 7,),
-	(TASK3NAME, FIREFLY,),
-	(TASK3UPTIMER, 20,),
-	(TASK4BOWLSCOUNT, 0,),
-	(TASK4COUNT, 0,),
-	(TASK4DOWNTIMER, 7,),
-	(TASK4NAME, PEAK,),
-	(TASK4UPTIMER, 25,),
-	(TEXTNAME, "biditi.txt",),
-	(UPSEC, 20,),
-]
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * start of WINDOW structures
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+
+WINDOWTUP = (
+	(ALPHA_CHANNEL, 1),  # Sets the opacity of the window. 0 = invisible 1 = completely visible. Values bewteen 0 & 1 will produce semi-transparent windows in SOME environments (The Raspberry Pi always has this value at 1 and cannot change.
+	(AUTO_CLOSE, False),  # If True, the window will automatically close itself
+	(AUTO_CLOSE_DURATION, 3),  # Number of seconds to wait before closing the window
+	(AUTO_SIZE_BUTTONS, None),  # True if Buttons in this Window should be sized to exactly fit the text on this.
+	(AUTO_SIZE_TEXT, None),  # True if Elements in Window should be sized to exactly fir the length of text
+	(BACKGROUND_COLOR, None),  # color of background
+	(BORDER_DEPTH, None),  # Default border depth (width) for all elements in the window
+	(BUTTON_COLOR, None),  # Default button colors for all buttons in the window
+	(DEBUGGER_ENABLED, True),  # Default border depth (width) for all elements in the window
+	(DEFAULT_BUTTON_ELEMENT_SIZE, (None, None)),  # (width, height) size in characters (wide) and rows (high) for all Button elements in this window
+	(DEFAULT_ELEMENT_SIZE, (45, 1)),  # size in characters (wide) and rows (high) for all elements in this window
+	(DISABLE_CLOSE, False),  # If True, the X button in the top right corner of the window will no work.  Use with caution and always give a way out toyour users
+	(DISABLE_MINIMIZE, False),  # if True the user won't be able to minimize window.  Good for taking over entire screen and staying that way.
+	(ELEMENT_JUSTIFICATION, "left"),  # All elements in the Window itself will have this justification 'left', 'right', 'center' are valid values
+	(ELEMENT_PADDING, None),  # Default amount of padding to put around elements in window (left/right, top/bottom) or ((left, right), (top, bottom))
+	(ENABLE_CLOSE_ATTEMPTED_EVENT, False),  # If True then the window will not close when 'X' clicked. Instead an event WINDOW_CLOSE_ATTEMPTED_EVENT if returned from window.read
+	(FINALIZE, False),  # If True then the Finalize method will be called. Use this rather than chaining .Finalize for cleaner code
+	(FONT, None),  # specifies the font family, size, etc
+	(FORCE_TOPLEVEL, False),  # If True will cause this window to skip the normal use of a hidden master window
+	(GRAB_ANYWHERE, False),  # If True can use mouse to click and drag to move the window. Almost every location of the window will work except input fields on some systems
+	(ICON, None),  # Can be either a filename or Base64 value. For Windows if filename, it MUST be ICO format. For Linux, must NOT be ICO
+	(KEEP_ON_TOP, False),  # If True, window will be created on top of all other windows on screen. It can be bumped down if another window created with this parm
+	(LAYOUT, None),  # The layout for the window. Can also be specified in the Layout method
+	(LOCATION, (None, None)),  # (x,y) location, in pixels, to locate the upper left corner of the window on the screen. Default is to center on screen.
+	(MARGINS, (None, None)),  # (left/right, top/bottom) Amount of pixels to leave inside the window's frame around the edges before your elements are shown.
+	(METADATA, None),  # User metadata that can be set to ANYTHING
+	(MODAL, False),  # If True then this window will be the only window a user can interact with until it is closed
+	(NO_TITLEBAR, False),  # If true, no titlebar nor frame will be shown on window. This means you cannot minimize the window and it will not show up on the taskbar
+	(PROGRESS_BAR_COLOR, (None, None)),  # (bar color, background color) Sets the default colors for all progress bars in the window
+	(RESIZABLE, False),  # If True, allows the user to resize the window. Note the not all Elements will change size or location when resizing.
+	(RETURN_KEYBOARD_EVENTS, False),  # if True key presses on the keyboard will be returned as Events from Read calls
+	(RIGHT_CLICK_MENU, None),  # A list of lists of Menu items to show when this element is right clicked. See user docs for exact format.
+	(RIGHT_CLICK_MENU_BACKGROUND_COLOR, None),  # Background color for right click menus
+	(RIGHT_CLICK_MENU_DISABLED_TEXT_COLOR, None),  # Text color for disabled right click menu items
+	(RIGHT_CLICK_MENU_FONT, None),  # Font for right click menus
+	(RIGHT_CLICK_MENU_SELECTED_COLORS, (None, None)),  # Text AND background colors for a selected item. Can be a Tuple OR a color string. simplified-button-color-string 'foreground on background'. Can be a single color if want to set only the background.
+	(RIGHT_CLICK_MENU_TEAROFF, False),  # If True then all right click menus can be torn off
+	(RIGHT_CLICK_MENU_TEXT_COLOR, None),  # Text color for right click menus
+	(SIZE, (None, None)),  # (width, height) size in pixels for this window. Normally the window is autosized to fit contents, not set to an absolute size by the user
+	(TEXT_JUSTIFICATION, None),  # Default text justification for all Text Elements in window
+	(TITLE, ""),  # The title that will be displayed in the Titlebar and on the Taskbar
+	(TITLEBAR_BACKGROUND_COLOR, None),  # If custom titlebar indicated by use_custom_titlebar, then use this as background color
+	(TITLEBAR_FONT, None),  # If custom titlebar indicated by use_custom_titlebar, then use this as title font
+	(TITLEBAR_ICON, None),  # If custom titlebar indicated by use_custom_titlebar, then use this as the icon (file or base64 bytes)
+	(TITLEBAR_TEXT_COLOR, None),  # If custom titlebar indicated by use_custom_titlebar, then use this as text color
+	(TRANSPARENT_COLOR, None),  # Any portion of the window that has this color will be completely transparent. You can even click through these spots to the window under this window.
+	(TTK_THEME, None),  # Set the tkinter ttk 'theme' of the window.  Default = DEFAULT_TTK_THEME.  Sets all ttk widgets to this theme as their default
+	(USE_CUSTOM_TITLEBAR, None),  # If True, then a custom titlebar will be used instead of the normal titlebar
+	(USE_DEFAULT_FOCUS, True),  # If True will use the default focus algorithm to set the focus to the 'Correct' element
+	(USE_TTK_BUTTONS, None),  # Affects all buttons in window. True = use ttk buttons. False = do not use ttk buttons.  None = use ttk buttons only if on a Mac
+)
+
+def WINDOWDICT():
+	return dict((x, y) for x, y in WINDOWTUP)
 
 
-def defaults():
-	defaultsRtn = {}
-	for entry in DEFAULTS:
-		defaultsRtn[entry[0]] = entry[1]
-	return defaultsRtn
+#
+#
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+# * end of managed sections of PSG.py
+# * #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*
+#
+#
 
 
-DEFAULTSVALUESNDX = {
-	TASK1NAME: 0,
-	TASK2NAME: 1,
-	TASK3NAME: 2,
-	TASK4NAME: 3,
+
+
+
+
+EMPTYALARM = {
+	ENABLED: True,
+	EVENTMODE: MODE_ALARM,
+	NAME: "",
+	PREDISMISSABLE: True,
+	SNOOZABLE: False,
+	TIMEALARM: "00:00:00",
 }
 
 
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# get a few things done that will be used by functions, but were unneeded by the init above
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-
-directionUp = True
-ticksPerSecond = TICKSPERSEC
-interruptRate = INTERRUPTRATE
-ticks = 0
-timerRunning = False
-currentData = None
-
-
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# buttons defined here, don't forget to ** double unpack these when used
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-
-ADJBTNDOWN = {
-	"button_text": "",
-	"font": (FONT, ADJTIMEBTNFONTSZ),
-	"image_data": ADJTIMEDOWNIMAGE,
-	"focus": True,
-	"button_color": (ADJBTNDOWNTEXTCOLOR, ADJBTNDOWNCOLOR),
-}
-
-ADJBTNUP = {
-	"button_text": "",
-	"image_data": ADJTIMEUPIMAGE,
-	"font": (FONT, ADJTIMEBTNFONTSZ),
-	"focus": True,
-	"button_color": (ADJBTNUPTEXTCOLOR, ADJBTNUPCOLOR),
-}
-
-BTNBOWLSDOWN = {
-	"button_color": (BTNDOWNTEXTCOLOR, BTNDOWNCOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BOWLSFONTSZ),
-	"image_data": BOWLSDOWN,
-}
-
-BTNBOWLSUP = {
-	"button_color": (BTNUPTEXTCOLOR, BTNUPCOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BOWLSFONTSZ),
-	"image_data": BOWLSUP,
-}
-
-BTNUP = {
-	"button_color": (BTNUPTEXTCOLOR, BTNUPCOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"image_data": UPIMAGE,
-}
-
-BTNQUIT = {
-	"button_color": (BTNQUITTEXTCOLOR, BTNQUITCOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"image_data": QUITBTN,
-}
-
-BTNRESETONE = {
-	"button_color": (BTNRESETONETEXTCOLOR, BTNRESETONECOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"image_data": RESETONE,
-}
-
-BTNTASKUP = {
-	"button_text": "",
-	"image_data": UPIMAGE,
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"button_color": (BTNTASKUPTEXTCOLOR, BTNTASKUPCOLOR),
-}
-
-BTNTASKDOWN = {
-	"button_color": (BTNTASKDOWNTEXTCOLOR, BTNTASKDOWNCOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"image_data": DOWNIMAGE,
-}
-
-BTNUPNOTIME = {
-	"button_color": (BTNUPNOTIMETEXTCOLOR, BTNUPNOTIMECOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"image_data": UPNOTIME,
-}
-
-BTNZEROALL = {
-	"button_color": (BTNZEROTEXTCOLOR, BTNZEROCOLOR),
-	"button_text": "",
-	"focus": True,
-	"font": (FONT, BTNFONTSZ),
-	"image_data": ZEROALL,
-}
-
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# spin box
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-
-
-TASKSPINBOX = {
-	"values": TASKLIST,
-	"background_color": SPINBKGNDCOLOR,
-	"font": (FONT, SPINFONTSZ,),
-	"size": SPINSIZE,
-	"text_color": SPINTEXTCOLOR,
+EMPTYALARMREMIND = {
+	ENABLED: True,
+	EVENTMODE: MODE_ALARM,
+	NAME: "",
+	PREDISMISSABLE: True,
+	SNOOZABLE: False,
+	TIMEALARM: "00:00:00",
+	TIMEREMIND: "00:00:00",
 }
 
 
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# text parameters
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-TEXTADJTIMEDOWN = {
-	"background_color": ADJTIMEDOWNBKGNDCOLOR,
-	"font": (FONT, ADJTIMEFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": ADJBTNDOWNTEXTCOLOR,
-}
-
-TEXTADJTIMEUP = {
-	"background_color": ADJTIMEUPBKGNDCOLOR,
-	"font": (FONT, ADJTIMEFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": ADJBTNUPTEXTCOLOR,
-}
-
-TEXTBOWLSCOUNT = {
-	# "background_color": TIMERDOWNBKGNDCOLOR,
-	"font": (FONT, BOWLSFONTSZ),
-	"justification": "center",
-	"size": (2, 1),
-	"text_color": TASKCOUNTERCOLOR,
-}
-
-TEXTTASKCOUNT = {
-	"font": (FONT, COUNTERFONTSZ),
-	"justification": "center",
-	"size": TASKCOUNTSZ,
-	"text_color": TASKCOUNTERCOLOR,
-}
-
-TEXTTIMERDOWN = {
-	"background_color": TIMERDOWNBKGNDCOLOR,
-	"font": (FONT, SETTIMERFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": TIMERDOWNTEXTCOLOR,
-}
-
-TEXTTIMERDOWNPARMS = {
-	"background_color": TIMERDOWNBKGNDCOLOR,
-	"text_color": TIMERDOWNTEXTCOLOR,
-}
-
-TEXTTIMEROFF = {
-	"background_color": TIMEROFFBKGNDCOLOR,
-	"font": (FONT, SETTIMERFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": TIMEROFFTXTCOLOR,
-}
-
-TEXTTIMEROFFPARMS = {
-	"background_color": TIMEROFFBKGNDCOLOR,
-	"text_color": TIMEROFFTXTCOLOR,
-}
-
-TEXTTIMERUP = {
-	"background_color": TIMERUPBKGNDCOLOR,
-	"font": (FONT, SETTIMERFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": TIMERUPTEXTCOLOR,
-}
-
-TEXTTIMERUPPARMS = {
-	"background_color": TIMERUPBKGNDCOLOR,
-	"text_color": TIMERUPTEXTCOLOR,
-}
-
-TEXTSPACE1 = {
-	"font": (FONT, SPACEFONTSZ),
-	"justification": "center",
-	"size": (3, 1),
-	"text_color": SPACECOLOR,
-}
-
-TEXTDOWNLABEL = {
-	"background_color": TIMERDOWNBKGNDCOLOR,
-	"size": (3, 1),
-	"text_color": TIMERDOWNTEXTCOLOR,
-}
-
-TEXTUPLABEL = {
-	"background_color": TIMERUPBKGNDCOLOR,
-	"size": (3, 1),
-	"text_color": TIMERUPTEXTCOLOR,
-}
-
-TEXTDOWNTIME = {
-	"background_color": TIMERDOWNBKGNDCOLOR,
-	"font": (FONT, UPDOWNTIMEFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": TIMERDOWNTEXTCOLOR,
-}
-
-TEXTUPTIME = {
-	"background_color": TIMERUPBKGNDCOLOR,
-	"font": (FONT, UPDOWNTIMEFONTSZ),
-	"justification": "center",
-	"size": (5, 1),
-	"text_color": TIMERUPTEXTCOLOR,
+EMPTYINTERVAL = {
+	ENABLED: True,
+	EVENTMODE: MODE_ALARM,
+	NAME: "",
+	TIMEINTERVAL: "00:00:00"
 }
 
 
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# columns, remember to ** unpack as appropriate
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-
-CLMTASK1 = [
-	[
-		SG.Spin(
-			initial_value=currentData[TASK1NAME],
-			key="_task1_spin_",
-			**TASKSPINBOX,
-		),
-	],
-	[
-		SG.Button(
-			key="task1+",
-			**BTNTASKUP,
-		),
-		SG.Btn(
-			key="task1+nt",
-			**BTNUPNOTIME,
-		),
-	],
-	[
-		SG.Text(
-			key="_task1count_",
-			**TEXTTASKCOUNT,
-		),
-	],
-	[
-		SG.Button(
-			key="task1-",
-			**BTNTASKDOWN,
-		),
-		SG.Btn(
-			key="task1reset",
-			**BTNRESETONE,
-		),
-	],
-	[
-		SG.Btn(
-			key="U1M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="U1S+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D1M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D1S+",
-			**ADJBTNUP,
-		),
-	],
-	[
-		SG.Text(
-			key="_task1UpTimer_",
-			**TEXTADJTIMEUP,
-		),
-		SG.Text(
-			key="_task1DownTimer_",
-			**TEXTADJTIMEDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="U1M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="U1S-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D1M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D1S-",
-			**ADJBTNDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="D1B",
-			**BTNBOWLSDOWN,
-		),
-		SG.Text(
-			key="_task1BowlsCount_",
-			**TEXTBOWLSCOUNT,
-		),
-		SG.Btn(
-			key="U1B",
-			**BTNBOWLSUP,
-		),
-	],
-]
-
-CLMTASK2 = [
-	[
-		SG.Spin(
-			initial_value=currentData[TASK2NAME],
-			key="_task2_spin_",
-			**TASKSPINBOX,
-		),
-	],
-	[
-		SG.Button(
-			key="task2+",
-			**BTNTASKUP,
-		),
-		SG.Btn(
-			key="task2+nt",
-			**BTNUPNOTIME,
-		),
-	],
-	[
-		SG.Text(
-			key="_task2count_",
-			**TEXTTASKCOUNT,
-		),
-	],
-	[
-		SG.Button(
-			key="task2-",
-			**BTNTASKDOWN,
-		),
-		SG.Btn(
-			key="task2reset",
-			**BTNRESETONE,
-		),
-	],
-	[
-		SG.Btn(
-			key="U2M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="U2S+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D2M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D2S+",
-			**ADJBTNUP,
-		),
-	],
-	[
-		SG.Text(
-			key="_task2UpTimer_",
-			**TEXTADJTIMEUP,
-		),
-		SG.Text(
-			key="_task2DownTimer_",
-			**TEXTADJTIMEDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="U2M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="U2S-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D2M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D2S-",
-			**ADJBTNDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="D2B",
-			**BTNBOWLSDOWN,
-		),
-		SG.Text(
-			key="_task2BowlsCount_",
-			**TEXTBOWLSCOUNT,
-		),
-		SG.Btn(
-			key="U2B",
-			**BTNBOWLSUP,
-		),
-	],
-]
-
-CLMTASK3 = [
-	[
-		SG.Spin(
-			initial_value=currentData[TASK3NAME],
-			key="_task3_spin_",
-			**TASKSPINBOX,
-		),
-	],
-	[
-		SG.Button(
-			key="task3+",
-			**BTNTASKUP,
-		),
-		SG.Btn(
-			key="task3+nt",
-			**BTNUPNOTIME,
-		),
-	],
-	[
-		SG.Text(
-			key="_task3count_",
-			**TEXTTASKCOUNT,
-		),
-	],
-	[
-		SG.Button(
-			key="task3-",
-			**BTNTASKDOWN,
-		),
-		SG.Btn(
-			key="task3reset",
-			**BTNRESETONE,
-		),
-	],
-	[
-		SG.Btn(
-			key="U3M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="U3S+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D3M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D3S+",
-			**ADJBTNUP,
-		),
-	],
-	[
-		SG.Text(
-			key="_task3UpTimer_",
-			**TEXTADJTIMEUP,
-		),
-		SG.Text(
-			key="_task3DownTimer_",
-			**TEXTADJTIMEDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="U3M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="U3S-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D3M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D3S-",
-			**ADJBTNDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="D3B",
-			**BTNBOWLSDOWN,
-		),
-		SG.Text(
-			key="_task3BowlsCount_",
-			**TEXTBOWLSCOUNT,
-		),
-		SG.Btn(
-			key="U3B",
-			**BTNBOWLSUP,
-		),
-	],
-]
-
-CLMTASK4 = [
-	[
-		SG.Spin(
-			initial_value=currentData[TASK4NAME],
-			key="_task4_spin_",
-			**TASKSPINBOX,
-		),
-	],
-	[
-		SG.Button(
-			key="task4+",
-			**BTNTASKUP,
-		),
-		SG.Btn(
-			key="task4+nt",
-			**BTNUPNOTIME,
-		),
-	],
-	[
-		SG.Text(
-			key="_task4count_",
-			**TEXTTASKCOUNT,
-		),
-	],
-	[
-		SG.Button(
-			key="task4-",
-			**BTNTASKDOWN,
-		),
-		SG.Btn(
-			key="task4reset",
-			**BTNRESETONE,
-		),
-	],
-	[
-		SG.Btn(
-			key="U4M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="U4S+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D4M+",
-			**ADJBTNUP,
-		),
-		SG.Btn(
-			key="D4S+",
-			**ADJBTNUP,
-		),
-	],
-	[
-		SG.Text(
-			key="_task4UpTimer_",
-			**TEXTADJTIMEUP,
-		),
-		SG.Text(
-			key="_task4DownTimer_",
-			**TEXTADJTIMEDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="U4M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="U4S-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D4M-",
-			**ADJBTNDOWN,
-		),
-		SG.Btn(
-			key="D4S-",
-			**ADJBTNDOWN,
-		),
-	],
-	[
-		SG.Btn(
-			key="D4B",
-			**BTNBOWLSDOWN,
-		),
-		SG.Text(
-			key="_task4BowlsCount_",
-			**TEXTBOWLSCOUNT,
-		),
-		SG.Btn(
-			key="U4B",
-			**BTNBOWLSUP,
-		),
-	],
-]
-
-CLMTIMER = [
-	[
-		SG.Text(
-			"000:00",
-			size=(6, 1),
-			text_color=TIMERUPTEXTCOLOR,
-			font=(FONT, TIMERFONTSZ),
-			justification="center",
-			key="_timer_",
-		),
-	],
-]
-
-mainLayout = [
-	[
-		SG.Col(CLMTIMER),
-	],
-	[
-		SG.Text(
-			key="_upTime_",
-			**TEXTUPTIME,
-		),
-		SG.Text(
-			key="_downTime_",
-			**TEXTDOWNTIME,
-		),
-		SG.Col(
-			[
-				[
-					SG.Btn(
-						key="Quit",
-						**BTNQUIT,
-					),
-				],
-				[
-					SG.Btn(
-						key="zeroAll",
-						**BTNZEROALL,
-					),
-				],
-			],
-		),
-	],
-	[
-		SG.Col(CLMTASK1),
-		SG.Col(CLMTASK2),
-		SG.Col(CLMTASK3),
-		SG.Col(CLMTASK4),
-	],
-]
-
-mainWindowParms = {
-	"title": MAINWINDOWTITLE,
-	"layout": mainLayout,
-	"location": (0, 0),
-	"element_padding": (0, 0),
-	"no_titlebar": False,
+EMPTYMAIN = {
+	CLOCKTIME: "00:00:00",
+	NEXTTIME: "00:00:00",
+	ELAPSEDTIME: "00:00:00",
+	TIMENEXTEVENT: "00:00:00",
+	INDEXNEXTEVENT: "00:00:00",
+	TRANSPARENT: True,
+	AVOIDMOUSE: True,
+	EVENTENTRIES: {
+		0: {
+			ENABLED: True,
+			EVENTMODE: MODE_ALARM,
+			NAME: "",
+			PREDISMISSABLE: True,
+			SNOOZABLE: False,
+			TIMEALARM: "04:00:00",
+		},
+		1: {
+			ENABLED: False,
+			EVENTMODE: MODE_ALARMREMIND,
+			NAME: "",
+			PREDISMISSABLE: True,
+			SNOOZABLE: False,
+			TIMEALARM: "00:00:00",
+			TIMEREMIND: "00:00:00",
+		},
+		2: {
+			TIMEALARM: "",
+			ENABLED: True,
+			EVENTMODE: MODE_INTERVAL,
+			TIMEINTERVAL: "00:00:00",
+			NAME: "",
+		},
+	},
 }
-
-mainFrame = SG.Window(
-	**mainWindowParms
-).finalize()
-
-
-"""Window(
-	title,
-	layout=None,
-	default_element_size=(45, 1),
-	default_button_element_size=(None, None),
-	auto_size_text=None,
-	auto_size_buttons=None,
-	location=(None, None),
-	size=(None, None),
-	element_padding=None,
-	margins=(None, None),
-	button_color=None,
-	font=None,
-	progress_bar_color=(None, None),
-	background_color=None,
-	border_depth=None,
-	auto_close=False,
-	auto_close_duration=3,
-	icon=None,
-	force_toplevel=False,
-	alpha_channel=1,
-	return_keyboard_events=False,
-	use_default_focus=True,
-	text_justification=None,
-	no_titlebar=False,
-	grab_anywhere=False,
-	keep_on_top=False,
-	resizable=False,
-	disable_close=False,
-	disable_minimize=False,
-	right_click_menu=None,
-	transparent_color=None,
-	debugger_enabled=True,
-	right_click_menu_background_color=None,
-	right_click_menu_text_color=None,
-	right_click_menu_disabled_text_color=None,
-	right_click_menu_selected_colors=(None, None),
-	right_click_menu_font=None,
-	right_click_menu_tearoff=False,
-	finalize=False,
-	element_justification='left',
-	ttk_theme=None,
-	use_ttk_buttons=None,
-	modal=False,
-	enable_close_attempted_event=False,
-	titlebar_background_color=None,
-	titlebar_text_color=None,
-	titlebar_font=None,
-	titlebar_icon=None,
-	use_custom_titlebar=None,
-	metadata=None)
-
-	Button(
-	button_text='',
-	button_type=7,
-	target=(None, None),
-	tooltip=None,
-	file_types=(('ALL Files', '*.*'),),
-	initial_folder=None,
-	default_extension='',
-	disabled=False,
-	change_submits=False,
-	enable_events=False,
-	image_filename=None,
-	image_data=None,
-	image_size=(None, None),
-	image_subsample=None,
-	border_width=None,
-	size=(None, None),
-	s=(None, None),
-	auto_size_button=None,
-	button_color=None,
-	disabled_button_color=None,
-	highlight_colors=None,
-	use_ttk_buttons=None,
-	font=None,
-	bind_return_key=False,
-	focus=False,
-	pad=None,
-	key=None,
-	k=None,
-	right_click_menu=None,
-	visible=True,
-	metadata=None)
-
-"""
-
-def updateAll():
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	# update timer and cycleCount
-	tempTimerVAL = ticks // ticksPerSecond
-	setTimer()
-	mainFrame.Element("_task1count_").Update(value=(f"{currentData[TASK1COUNT]:03d}"))
-	mainFrame.Element("_task2count_").Update(value=(f"{currentData[TASK2COUNT]:03d}"))
-	mainFrame.Element("_task3count_").Update(value=(f"{currentData[TASK3COUNT]:03d}"))
-	mainFrame.Element("_task4count_").Update(value=(f"{currentData[TASK4COUNT]:03d}"))
-
-	mainFrame.Element("_task1BowlsCount_").Update(value=(f"{currentData[TASK1BOWLSCOUNT]:02d}"))
-	mainFrame.Element("_task2BowlsCount_").Update(value=(f"{currentData[TASK2BOWLSCOUNT]:02d}"))
-	mainFrame.Element("_task3BowlsCount_").Update(value=(f"{currentData[TASK3BOWLSCOUNT]:02d}"))
-	mainFrame.Element("_task4BowlsCount_").Update(value=(f"{currentData[TASK4BOWLSCOUNT]:02d}"))
-
-	TUpTimer = makeTime(currentData[TASK1UPTIMER])
-	TDownTimer = makeTime(currentData[TASK1DOWNTIMER])
-	mainFrame.Element("_task1UpTimer_").Update(value=f"{TUpTimer}")
-	mainFrame.Element("_task1DownTimer_").Update(value=f"{TDownTimer}")
-
-	TUpTimer = makeTime(currentData[TASK2UPTIMER])
-	TDownTimer = makeTime(currentData[TASK2DOWNTIMER])
-	mainFrame.Element("_task2UpTimer_").Update(value=f"{TUpTimer}")
-	mainFrame.Element("_task2DownTimer_").Update(value=f"{TDownTimer}")
-
-	TUpTimer = makeTime(currentData[TASK3UPTIMER])
-	TDownTimer = makeTime(currentData[TASK3DOWNTIMER])
-	mainFrame.Element("_task3UpTimer_").Update(value=f"{TUpTimer}")
-	mainFrame.Element("_task3DownTimer_").Update(value=f"{TDownTimer}")
-
-	TUpTimer = makeTime(currentData[TASK4UPTIMER])
-	TDownTimer = makeTime(currentData[TASK4DOWNTIMER])
-	mainFrame.Element("_task4UpTimer_").Update(value=f"{TUpTimer}")
-	mainFrame.Element("_task4DownTimer_").Update(value=f"{TDownTimer}")
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def incCount(_inCount_):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	_T1_ = _inCount_ + 1
-	if _T1_ > 999:
-		_T1_ = 999
-	return _T1_
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def incBowlCount(_inCount_):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	_T1_ = _inCount_ + 1
-	if _T1_ > 99:
-		_T1_ = 99
-	return _T1_
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def decCount(_inCount_):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	TI = _inCount_ - 1
-	if TI < 0:
-		TI = 0
-	return TI
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def decBowlCount(_inCount_):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	_T1_ = _inCount_ - 1
-	if _T1_ < 0:
-		_T1_ = 0
-	return _T1_
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def incTime(inSeconds, increment):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	TInSeconds = inSeconds + increment
-	TMin = TInSeconds // 60
-	# TSec = TInSeconds % 60
-	if TMin > 999:
-		TInSeconds = (999 * 60) + 59
-	return TInSeconds
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def decTime(inSeconds, increment):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	TInSeconds = inSeconds - increment
-	if TInSeconds < 0:
-		TInSeconds = 0
-	return TInSeconds
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def updateMainBackground(COLOR):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	# put change background code
-	mainFrame.Element("_timer_").Update(background_color=COLOR)
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def setTimerOff():
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global ticks, directionUp
-	directionUp = True
-	ticks = 0
-	mainFrame.Element("_timer_").Update(**TEXTTIMEROFFPARMS)
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def setTimerUp():
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global directionUp, ticks
-	directionUp = True
-	ticks = 0
-	mainFrame.Element("_timer_").Update(**TEXTTIMERUPPARMS)
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def zeroStuff():
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global ticks, directionUp, currentData, timerRunning
-	ticks = 0
-	timerRunning = False
-	directionUp = True
-	currentData[EVENTS] = []
-	currentData[TASK1COUNT] = 0
-	currentData[TASK2COUNT] = 0
-	currentData[TASK3COUNT] = 0
-	currentData[TASK4COUNT] = 0
-	currentData[TASK1BOWLSCOUNT] = 0
-	currentData[TASK2BOWLSCOUNT] = 0
-	currentData[TASK3BOWLSCOUNT] = 0
-	currentData[TASK4BOWLSCOUNT] = 0
-	pickleIt(currentData[FILENAME], currentData)
-	updateAll()
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def startTimer():
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global timerRunning, currentData, ticks, directionUp
-	timerRunning = True
-	ticks = 0
-	directionUp = True
-	mainFrame.Element("_timer_").Update(**TEXTTIMERUPPARMS)
-	setTimer()
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def stopTimer():
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global timerRunning, ticks
-	timerRunning = False
-	setTimerOff()
-	updateAll()
-	mainFrame.Element("_timer_").Update(**TEXTTIMEROFFPARMS)
-	ticks = 0
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def getValues(inValues):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global currentData
-	for key, item in DEFAULTSVALUESNDX.items():
-		currentData[key] = inValues[item]
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def addEvent(event2add):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global currentData
-	return
-	entryToAdd = [nowStr(DT.now()), event2add]
-	# currentData[EVENTS].append(entryToAdd)
-	pickleIt(currentData[FILENAME], currentData)
-	with open(CONFIGDIRECTORY + currentData[TEXTNAME], "ta") as FDOut:
-		outStr = ""
-		outStr += f"""{entryToAdd}	{currentData[TASK1COUNT]}	{currentData[TASK2COUNT]}	{currentData[TASK3COUNT]}	{currentData[TASK4COUNT]}
-"""
-		# print(outStr)
-		FDOut.writelines(outStr)
-		FDOut.flush()
-		FDOut.close()
-	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-
-def doEvents(event):
-	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
-	global currentData
-	if timerRunning is True:
-		return
-	if event == "zeroAll":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		zeroStuff()
-		stopTimer()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "task1+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1COUNT] = incCount(currentData[TASK1COUNT])
-		currentData[UPSEC] = currentData[TASK1UPTIMER]
-		currentData[DOWNSEC] = currentData[TASK1DOWNTIMER]
-		updateAll()
-		startTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task2+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2COUNT] = incCount(currentData[TASK2COUNT])
-		currentData[UPSEC] = currentData[TASK2UPTIMER]
-		currentData[DOWNSEC] = currentData[TASK2DOWNTIMER]
-		updateAll()
-		startTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task3+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3COUNT] = incCount(currentData[TASK3COUNT])
-		currentData[UPSEC] = currentData[TASK3UPTIMER]
-		currentData[DOWNSEC] = currentData[TASK3DOWNTIMER]
-		updateAll()
-		startTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task4+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4COUNT] = incCount(currentData[TASK4COUNT])
-		currentData[UPSEC] = currentData[TASK4UPTIMER]
-		currentData[DOWNSEC] = currentData[TASK4DOWNTIMER]
-		updateAll()
-		startTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "task1-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1COUNT] = decCount(currentData[TASK1COUNT])
-		updateAll()
-		stopTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task2-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2COUNT] = decCount(currentData[TASK2COUNT])
-		updateAll()
-		stopTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task3-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3COUNT] = decCount(currentData[TASK3COUNT])
-		updateAll()
-		stopTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task4-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4COUNT] = decCount(currentData[TASK4COUNT])
-		updateAll()
-		stopTimer()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U1M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1UPTIMER] = incTime(currentData[TASK1UPTIMER], 60)
-		updateAll()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U1S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1UPTIMER] = incTime(currentData[TASK1UPTIMER], 1)
-		updateAll()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D1M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1DOWNTIMER] = incTime(currentData[TASK1DOWNTIMER], 60)
-		updateAll()
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D1S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1DOWNTIMER] = incTime(currentData[TASK1DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U2M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2UPTIMER] = incTime(currentData[TASK2UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U2S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2UPTIMER] = incTime(currentData[TASK2UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D2M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2DOWNTIMER] = incTime(currentData[TASK2DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D2S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2DOWNTIMER] = incTime(currentData[TASK2DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U3M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3UPTIMER] = incTime(currentData[TASK3UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U3S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3UPTIMER] = incTime(currentData[TASK3UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D3M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3DOWNTIMER] = incTime(currentData[TASK3DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D3S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3DOWNTIMER] = incTime(currentData[TASK3DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U4M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4UPTIMER] = incTime(currentData[TASK4UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U4S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4UPTIMER] = incTime(currentData[TASK4UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D4M+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4DOWNTIMER] = incTime(currentData[TASK4DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D4S+":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4DOWNTIMER] = incTime(currentData[TASK4DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U1M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1UPTIMER] = decTime(currentData[TASK1UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U1S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1UPTIMER] = decTime(currentData[TASK1UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D1M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1DOWNTIMER] = decTime(currentData[TASK1DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D1S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1DOWNTIMER] = decTime(currentData[TASK1DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U2M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2UPTIMER] = decTime(currentData[TASK2UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U2S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2UPTIMER] = decTime(currentData[TASK2UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D2M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2DOWNTIMER] = decTime(currentData[TASK2DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D2S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2DOWNTIMER] = decTime(currentData[TASK2DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U3M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3UPTIMER] = decTime(currentData[TASK3UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U3S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3UPTIMER] = decTime(currentData[TASK3UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D3M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3DOWNTIMER] = decTime(currentData[TASK3DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D3S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3DOWNTIMER] = decTime(currentData[TASK3DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U4M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4UPTIMER] = decTime(currentData[TASK4UPTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U4S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4UPTIMER] = decTime(currentData[TASK4UPTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D4M-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4DOWNTIMER] = decTime(currentData[TASK4DOWNTIMER], 60)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D4S-":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4DOWNTIMER] = decTime(currentData[TASK4DOWNTIMER], 1)
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "task1+nt":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1COUNT] = incCount(currentData[TASK1COUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task2+nt":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2COUNT] = incCount(currentData[TASK2COUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task3+nt":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3COUNT] = incCount(currentData[TASK3COUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task4+nt":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4COUNT] = incCount(currentData[TASK4COUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "task1reset":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1COUNT] = 0
-		currentData[TASK1BOWLSCOUNT] = 0
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task2reset":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2COUNT] = 0
-		currentData[TASK2BOWLSCOUNT] = 0
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task3reset":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3COUNT] = 0
-		currentData[TASK3BOWLSCOUNT] = 0
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "task4reset":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4COUNT] = 0
-		currentData[TASK4BOWLSCOUNT] = 0
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "D1B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1BOWLSCOUNT] = decBowlCount(currentData[TASK1BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D2B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2BOWLSCOUNT] = decBowlCount(currentData[TASK2BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D3B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3BOWLSCOUNT] = decBowlCount(currentData[TASK3BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "D4B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4BOWLSCOUNT] = decBowlCount(currentData[TASK4BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-
-	elif event == "U1B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK1BOWLSCOUNT] = incBowlCount(currentData[TASK1BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U2B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK2BOWLSCOUNT] = incBowlCount(currentData[TASK2BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U3B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK3BOWLSCOUNT] = incBowlCount(currentData[TASK3BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "U4B":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		currentData[TASK4BOWLSCOUNT] = incBowlCount(currentData[TASK4BOWLSCOUNT])
-		updateAll()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
-
-getData(currentData[FILENAME])
-updateAll()
-stopTimer()
-while True:  # Event Loop
-	event, values = mainFrame.Read(timeout=interruptRate)  # use as high of a timeout value as you can
-	# 	print(f"""
-	# event |{event}|
-	# values |{values}|
-	# """)
-	if (event is None) or (event == "Quit"):  # X or quit button clicked
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		addEvent(event)
-		stopTimer()
-		pickleIt(currentData[FILENAME], currentData)
-		break
-	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	elif event == "__TIMEOUT__":
-		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
-		if directionUp is True:
-			ticks += 1
-		else:
-			ticks -= 1
-		if abs(ticks) > MAXTICKS:
-			ticks = 0
-		# print(f"ticks {ticks}")
-		if timerRunning is True:
-			upTicks = int(currentData[UPSEC] * ticksPerSecond)
-			downTicks = int(currentData[DOWNSEC] * ticksPerSecond)
-			if (directionUp is True) and (ticks > upTicks):
-				mainFrame.Element("_timer_").Update(**TEXTTIMERDOWNPARMS)
-				directionUp = False
-				ticks = downTicks
-			elif (directionUp is False) & (ticks <= ticksPerSecond):
-				stopTimer()
-		setTimer()
-		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
-	else:
-		doEvents(event)
-
-
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# end of while True
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-
-
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
-# end of biditi.proper
-# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
