+++ biditi.py
@@ -1,0 +1,1560 @@
+#!/usr/bin/python
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# import globally
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+from datetime import datetime as DT
+from os import path as PATH
+import pickle as PD
+import PySimpleGUI as SG
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# setting constants
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+BLACK = "#000000"
+GRAY8 = "#888888"
+GRAYC = "#CCCCCC"
+GREEN2 = "#448811"
+MAINDOWNCOLOR = "#880000"
+MAINDOWNTEXTCOLOR = "#FF0000"
+MAINUPCOLOR = "#008800"
+MAINUPTEXTCOLOR = "#00FF00"
+ORANGE = "#FF7F00"
+PURP440022 = "#440022"
+PURP660044 = "#660044"
+TEAL1 = "#00FF7F"
+YELLOW2 = "#444422"
+YELLOW666600 = "#666600"
+
+
+AIR = "AIR"
+DYNAVAP = "DYNAVAP"
+FIREFLY = "FIREFLY"
+PEAK = "PEAK"
+Q = "Q"
+QDDM = "QDDM"
+QOMO = "QOMO"
+QTZBANGER = "QTZBANGER"
+VARIOUS = "VARIOUS"
+
+
+ADJBTNDOWNCOLOR = MAINDOWNCOLOR
+ADJBTNDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
+ADJBTNUPCOLOR = MAINUPCOLOR
+ADJBTNUPTEXTCOLOR = MAINUPTEXTCOLOR
+ADJTIMEDOWNBKGNDCOLOR = MAINDOWNCOLOR
+ADJTIMETXTCOLOR = GRAYC
+ADJTIMEUPBKGNDCOLOR = MAINUPCOLOR
+BTNDOWNCOLOR = MAINDOWNCOLOR
+BTNDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
+BTNQUITCOLOR = PURP440022
+BTNQUITTEXTCOLOR = YELLOW666600
+BTNRESETONECOLOR = MAINDOWNCOLOR
+BTNRESETONETEXTCOLOR = MAINDOWNTEXTCOLOR
+BTNTASKDOWNCOLOR = MAINDOWNCOLOR
+BTNTASKDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
+BTNTASKUPCOLOR = MAINUPCOLOR
+BTNTASKUPTEXTCOLOR = MAINUPTEXTCOLOR
+BTNUPCOLOR = MAINUPCOLOR
+BTNUPNOTIMECOLOR = MAINUPCOLOR
+BTNUPNOTIMETEXTCOLOR = MAINUPTEXTCOLOR
+BTNUPTEXTCOLOR = MAINUPTEXTCOLOR
+BTNZEROCOLOR = PURP440022
+BTNZEROTEXTCOLOR = YELLOW666600
+INTERRUPTRATE = 100
+LASTFILENAME = "biditi.last"
+MAXTICKS = 59999
+MODE_NORMAL = "MODE_NORMAL"
+MODE_RESTART = "MODE_RESTART"
+MODE_START = "MODE_START"
+NOT = "NOT"
+SPACECOLOR = GRAY8
+SPINBKGNDCOLOR = TEAL1
+SPINSIZE = (10, 1)
+SPINTEXTCOLOR = ORANGE
+TASKCOUNTERCOLOR = GREEN2
+TASKCOUNTSZ = (3, 1)
+TICKSPERSEC = 10
+TIMERDOWNBKGNDCOLOR = MAINDOWNCOLOR
+TIMERDOWNTEXTCOLOR = MAINDOWNTEXTCOLOR
+TIMEROFFBKGNDCOLOR = BLACK
+TIMEROFFTXTCOLOR = YELLOW2
+TIMERUPBKGNDCOLOR = MAINUPCOLOR
+TIMERUPTEXTCOLOR = MAINUPTEXTCOLOR
+
+
+PCADJTIMEBTNFONTSZ = 6
+PCADJTIMEDOWNIMAGE = DN32
+PCADJTIMEFONTSZ = 18
+PCADJTIMEUPIMAGE = UP32
+PCBOWLSDOWN = BDOWN32
+PCBOWLSFONTSZ = 18
+PCBOWLSUP = BUP32
+PCBTNFONTSZ = 10
+PCCOUNTERFONTSZ = 30
+PCDOWNIMAGE = DN64
+PCFONT = "Source Code Pro"
+PCLABELFONTSZ = 12
+PCQUITBTN = QUIT64
+PCRESETONE = RESETONE64
+PCSETTIMERFONTSZ = 20
+PCSPACEFONTSZ = 10
+PCSPINFONTSZ = 16
+PCTIMERFONTSZ = 130
+PCUPDOWNTIMEFONTSZ = 50
+PCUPIMAGE = UP64
+PCUPNOTIME = UPNOTIME64
+PCZEROALL = ZEROALL64
+
+
+CWD = PATH.abspath(".")
+if CWD.find("_android") > -1:
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	ADJTIMEBTNFONTSZ = 6
+	ADJTIMEDOWNIMAGE = DN64
+	ADJTIMEFONTSZ = 10
+	ADJTIMEUPIMAGE = UP64
+	BOWLSDOWN = BDOWN64
+	BOWLSFONTSZ = 10
+	BOWLSUP = BUP64
+	BTNFONTSZ = 12
+	CONFIGDIRECTORY = ""
+	COUNTERFONTSZ = 20
+	DOWNIMAGE = DN128
+	FONT = "Source Code Pro"
+	LABELFONTSZ = 12
+	QUITBTN = QUIT128
+	RESETONE = RESETONE128
+	SETTIMERFONTSZ = 20
+	SG.ChangeLookAndFeel("DarkPurple6")
+	SPACEFONTSZ = 10
+	SPINFONTSZ = 9
+	TIMERFONTSZ = 70
+	UPDOWNTIMEFONTSZ = 34
+	UPIMAGE = UP128
+	UPNOTIME = UPNOTIME128
+	ZEROALL = ZEROALL128
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+elif CWD.find("_DEV") > -1:
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	ADJTIMEBTNFONTSZ = PCADJTIMEBTNFONTSZ
+	ADJTIMEDOWNIMAGE = PCADJTIMEDOWNIMAGE
+	ADJTIMEFONTSZ = PCADJTIMEFONTSZ
+	ADJTIMEUPIMAGE = PCADJTIMEUPIMAGE
+	BOWLSDOWN = PCBOWLSDOWN
+	BOWLSFONTSZ = PCBOWLSFONTSZ
+	BOWLSUP = PCBOWLSUP
+	BTNFONTSZ = PCBTNFONTSZ
+	CONFIGDIRECTORY = "/home/will/.config/biditi_DEV/"
+	COUNTERFONTSZ = PCCOUNTERFONTSZ
+	DOWNIMAGE = PCDOWNIMAGE
+	FONT = PCFONT
+	LABELFONTSZ = PCLABELFONTSZ
+	QUITBTN = PCQUITBTN
+	RESETONE = PCRESETONE
+	SETTIMERFONTSZ = PCSETTIMERFONTSZ
+	SG.ChangeLookAndFeel("DarkGreen5")
+	SPACEFONTSZ = PCSPACEFONTSZ
+	SPINFONTSZ = PCSPINFONTSZ
+	TIMERFONTSZ = PCTIMERFONTSZ
+	UPDOWNTIMEFONTSZ = PCUPDOWNTIMEFONTSZ
+	UPIMAGE = PCUPIMAGE
+	UPNOTIME = PCUPNOTIME
+	ZEROALL = PCZEROALL
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+else:
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	ADJTIMEBTNFONTSZ = PCADJTIMEBTNFONTSZ
+	ADJTIMEDOWNIMAGE = PCADJTIMEDOWNIMAGE
+	ADJTIMEFONTSZ = PCADJTIMEFONTSZ
+	ADJTIMEUPIMAGE = PCADJTIMEUPIMAGE
+	BOWLSDOWN = PCBOWLSDOWN
+	BOWLSFONTSZ = PCBOWLSFONTSZ
+	BOWLSUP = PCBOWLSUP
+	BTNFONTSZ = PCBTNFONTSZ
+	CONFIGDIRECTORY = "/home/will/.config/biditi/"
+	COUNTERFONTSZ = PCCOUNTERFONTSZ
+	DOWNIMAGE = PCDOWNIMAGE
+	FONT = PCFONT
+	LABELFONTSZ = PCLABELFONTSZ
+	QUITBTN = PCQUITBTN
+	RESETONE = PCRESETONE
+	SETTIMERFONTSZ = PCSETTIMERFONTSZ
+	SG.ChangeLookAndFeel("DarkPurple6")
+	SPACEFONTSZ = PCSPACEFONTSZ
+	SPINFONTSZ = PCSPINFONTSZ
+	TIMERFONTSZ = PCTIMERFONTSZ
+	UPDOWNTIMEFONTSZ = PCUPDOWNTIMEFONTSZ
+	UPIMAGE = PCUPIMAGE
+	UPNOTIME = PCUPNOTIME
+	ZEROALL = PCZEROALL
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+if CWD.find("_DEV") > -1:
+	DEVMODE = True
+	MAINWINDOWTITLE = "BIDITI_DEV"
+else:
+	DEVMODE = False
+	MAINWINDOWTITLE = "biditi"
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# currentData dict keys and uses
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+BUTTON = "BUTTON"
+DOWNSEC = "DOWNSEC"
+EVENTS = "EVENTS"
+FILENAME = "FILENAME"
+TASK1BOWLSCOUNT = "TASK1BOWLSCOUNT"
+TASK1COUNT = "TASK1COUNT"
+TASK1DOWNTIMER = "TASK1DOWNTIMER"
+TASK1NAME = "TASK1NAME"
+TASK1UPTIMER = "TASK1UPTIMER"
+TASK2BOWLSCOUNT = "TASK2BOWLSCOUNT"
+TASK2COUNT = "TASK2COUNT"
+TASK2DOWNTIMER = "TASK2DOWNTIMER"
+TASK2NAME = "TASK2NAME"
+TASK2UPTIMER = "TASK2UPTIMER"
+TASK3BOWLSCOUNT = "TASK3BOWLSCOUNT"
+TASK3COUNT = "TASK3COUNT"
+TASK3DOWNTIMER = "TASK3DOWNTIMER"
+TASK3NAME = "TASK3NAME"
+TASK3UPTIMER = "TASK3UPTIMER"
+TASK4BOWLSCOUNT = "TASK4BOWLSCOUNT"
+TASK4COUNT = "TASK4COUNT"
+TASK4DOWNTIMER = "TASK4DOWNTIMER"
+TASK4NAME = "TASK4NAME"
+TASK4UPTIMER = "TASK4UPTIMER"
+TEXTNAME = "TEXTNAME"
+UPSEC = "UPSEC"
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# the defaults for all things settable via currentData
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+DEFAULTS = [
+	(DOWNSEC, 7,),
+	(EVENTS, [],),
+	(FILENAME, "biditi.pkl",),
+	(TASK1BOWLSCOUNT, 0,),
+	(TASK1COUNT, 0,),
+	(TASK1DOWNTIMER, 7,),
+	(TASK1NAME, Q,),
+	(TASK1UPTIMER, 20,),
+	(TASK2BOWLSCOUNT, 0,),
+	(TASK2COUNT, 0,),
+	(TASK2DOWNTIMER, 7,),
+	(TASK2NAME, QTZBANGER,),
+	(TASK2UPTIMER, 20,),
+	(TASK3BOWLSCOUNT, 0,),
+	(TASK3COUNT, 0,),
+	(TASK3DOWNTIMER, 7,),
+	(TASK3NAME, FIREFLY,),
+	(TASK3UPTIMER, 20,),
+	(TASK4BOWLSCOUNT, 0,),
+	(TASK4COUNT, 0,),
+	(TASK4DOWNTIMER, 7,),
+	(TASK4NAME, PEAK,),
+	(TASK4UPTIMER, 25,),
+	(TEXTNAME, "biditi.txt",),
+	(UPSEC, 20,),
+]
+
+
+def defaults():
+	defaultsRtn = {}
+	for entry in DEFAULTS:
+		defaultsRtn[entry[0]] = entry[1]
+	return defaultsRtn
+
+
+DEFAULTSVALUESNDX = {
+	TASK1NAME: 0,
+	TASK2NAME: 1,
+	TASK3NAME: 2,
+	TASK4NAME: 3,
+}
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# get a few things done that will be used by functions, but were unneeded by the init above
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+directionUp = True
+ticksPerSecond = TICKSPERSEC
+interruptRate = INTERRUPTRATE
+ticks = 0
+timerRunning = False
+currentData = None
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# buttons defined here, don't forget to ** double unpack these when used
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+ADJBTNDOWN = {
+	"button_text": "",
+	"font": (FONT, ADJTIMEBTNFONTSZ),
+	"image_data": ADJTIMEDOWNIMAGE,
+	"focus": True,
+	"button_color": (ADJBTNDOWNTEXTCOLOR, ADJBTNDOWNCOLOR),
+}
+
+ADJBTNUP = {
+	"button_text": "",
+	"image_data": ADJTIMEUPIMAGE,
+	"font": (FONT, ADJTIMEBTNFONTSZ),
+	"focus": True,
+	"button_color": (ADJBTNUPTEXTCOLOR, ADJBTNUPCOLOR),
+}
+
+BTNBOWLSDOWN = {
+	"button_color": (BTNDOWNTEXTCOLOR, BTNDOWNCOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BOWLSFONTSZ),
+	"image_data": BOWLSDOWN,
+}
+
+BTNBOWLSUP = {
+	"button_color": (BTNUPTEXTCOLOR, BTNUPCOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BOWLSFONTSZ),
+	"image_data": BOWLSUP,
+}
+
+BTNUP = {
+	"button_color": (BTNUPTEXTCOLOR, BTNUPCOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"image_data": UPIMAGE,
+}
+
+BTNQUIT = {
+	"button_color": (BTNQUITTEXTCOLOR, BTNQUITCOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"image_data": QUITBTN,
+}
+
+BTNRESETONE = {
+	"button_color": (BTNRESETONETEXTCOLOR, BTNRESETONECOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"image_data": RESETONE,
+}
+
+BTNTASKUP = {
+	"button_text": "",
+	"image_data": UPIMAGE,
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"button_color": (BTNTASKUPTEXTCOLOR, BTNTASKUPCOLOR),
+}
+
+BTNTASKDOWN = {
+	"button_color": (BTNTASKDOWNTEXTCOLOR, BTNTASKDOWNCOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"image_data": DOWNIMAGE,
+}
+
+BTNUPNOTIME = {
+	"button_color": (BTNUPNOTIMETEXTCOLOR, BTNUPNOTIMECOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"image_data": UPNOTIME,
+}
+
+BTNZEROALL = {
+	"button_color": (BTNZEROTEXTCOLOR, BTNZEROCOLOR),
+	"button_text": "",
+	"focus": True,
+	"font": (FONT, BTNFONTSZ),
+	"image_data": ZEROALL,
+}
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# spin box
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+TASKLIST = [
+	AIR,
+	DYNAVAP,
+	FIREFLY,
+	PEAK,
+	Q,
+	QDDM,
+	QOMO,
+	QTZBANGER,
+	VARIOUS,
+]
+
+
+TASKSPINBOX = {
+	"values": TASKLIST,
+	"background_color": SPINBKGNDCOLOR,
+	"font": (FONT, SPINFONTSZ,),
+	"size": SPINSIZE,
+	"text_color": SPINTEXTCOLOR,
+}
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# text parameters
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+TEXTADJTIMEDOWN = {
+	"background_color": ADJTIMEDOWNBKGNDCOLOR,
+	"font": (FONT, ADJTIMEFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": ADJBTNDOWNTEXTCOLOR,
+}
+
+TEXTADJTIMEUP = {
+	"background_color": ADJTIMEUPBKGNDCOLOR,
+	"font": (FONT, ADJTIMEFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": ADJBTNUPTEXTCOLOR,
+}
+
+TEXTBOWLSCOUNT = {
+	# "background_color": TIMERDOWNBKGNDCOLOR,
+	"font": (FONT, BOWLSFONTSZ),
+	"justification": "center",
+	"size": (2, 1),
+	"text_color": TASKCOUNTERCOLOR,
+}
+
+TEXTTASKCOUNT = {
+	"font": (FONT, COUNTERFONTSZ),
+	"justification": "center",
+	"size": TASKCOUNTSZ,
+	"text_color": TASKCOUNTERCOLOR,
+}
+
+TEXTTIMERDOWN = {
+	"background_color": TIMERDOWNBKGNDCOLOR,
+	"font": (FONT, SETTIMERFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": TIMERDOWNTEXTCOLOR,
+}
+
+TEXTTIMERDOWNPARMS = {
+	"background_color": TIMERDOWNBKGNDCOLOR,
+	"text_color": TIMERDOWNTEXTCOLOR,
+}
+
+TEXTTIMEROFF = {
+	"background_color": TIMEROFFBKGNDCOLOR,
+	"font": (FONT, SETTIMERFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": TIMEROFFTXTCOLOR,
+}
+
+TEXTTIMEROFFPARMS = {
+	"background_color": TIMEROFFBKGNDCOLOR,
+	"text_color": TIMEROFFTXTCOLOR,
+}
+
+TEXTTIMERUP = {
+	"background_color": TIMERUPBKGNDCOLOR,
+	"font": (FONT, SETTIMERFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": TIMERUPTEXTCOLOR,
+}
+
+TEXTTIMERUPPARMS = {
+	"background_color": TIMERUPBKGNDCOLOR,
+	"text_color": TIMERUPTEXTCOLOR,
+}
+
+TEXTSPACE1 = {
+	"font": (FONT, SPACEFONTSZ),
+	"justification": "center",
+	"size": (3, 1),
+	"text_color": SPACECOLOR,
+}
+
+TEXTDOWNLABEL = {
+	"background_color": TIMERDOWNBKGNDCOLOR,
+	"size": (3, 1),
+	"text_color": TIMERDOWNTEXTCOLOR,
+}
+
+TEXTUPLABEL = {
+	"background_color": TIMERUPBKGNDCOLOR,
+	"size": (3, 1),
+	"text_color": TIMERUPTEXTCOLOR,
+}
+
+TEXTDOWNTIME = {
+	"background_color": TIMERDOWNBKGNDCOLOR,
+	"font": (FONT, UPDOWNTIMEFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": TIMERDOWNTEXTCOLOR,
+}
+
+TEXTUPTIME = {
+	"background_color": TIMERUPBKGNDCOLOR,
+	"font": (FONT, UPDOWNTIMEFONTSZ),
+	"justification": "center",
+	"size": (5, 1),
+	"text_color": TIMERUPTEXTCOLOR,
+}
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# columns, remember to ** unpack as appropriate
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+CLMTASK1 = [
+	[
+		SG.Spin(
+			initial_value=currentData[TASK1NAME],
+			key="_task1_spin_",
+			**TASKSPINBOX,
+		),
+	],
+	[
+		SG.Button(
+			key="task1+",
+			**BTNTASKUP,
+		),
+		SG.Btn(
+			key="task1+nt",
+			**BTNUPNOTIME,
+		),
+	],
+	[
+		SG.Text(
+			key="_task1count_",
+			**TEXTTASKCOUNT,
+		),
+	],
+	[
+		SG.Button(
+			key="task1-",
+			**BTNTASKDOWN,
+		),
+		SG.Btn(
+			key="task1reset",
+			**BTNRESETONE,
+		),
+	],
+	[
+		SG.Btn(
+			key="U1M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="U1S+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D1M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D1S+",
+			**ADJBTNUP,
+		),
+	],
+	[
+		SG.Text(
+			key="_task1UpTimer_",
+			**TEXTADJTIMEUP,
+		),
+		SG.Text(
+			key="_task1DownTimer_",
+			**TEXTADJTIMEDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="U1M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="U1S-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D1M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D1S-",
+			**ADJBTNDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="D1B",
+			**BTNBOWLSDOWN,
+		),
+		SG.Text(
+			key="_task1BowlsCount_",
+			**TEXTBOWLSCOUNT,
+		),
+		SG.Btn(
+			key="U1B",
+			**BTNBOWLSUP,
+		),
+	],
+]
+
+CLMTASK2 = [
+	[
+		SG.Spin(
+			initial_value=currentData[TASK2NAME],
+			key="_task2_spin_",
+			**TASKSPINBOX,
+		),
+	],
+	[
+		SG.Button(
+			key="task2+",
+			**BTNTASKUP,
+		),
+		SG.Btn(
+			key="task2+nt",
+			**BTNUPNOTIME,
+		),
+	],
+	[
+		SG.Text(
+			key="_task2count_",
+			**TEXTTASKCOUNT,
+		),
+	],
+	[
+		SG.Button(
+			key="task2-",
+			**BTNTASKDOWN,
+		),
+		SG.Btn(
+			key="task2reset",
+			**BTNRESETONE,
+		),
+	],
+	[
+		SG.Btn(
+			key="U2M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="U2S+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D2M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D2S+",
+			**ADJBTNUP,
+		),
+	],
+	[
+		SG.Text(
+			key="_task2UpTimer_",
+			**TEXTADJTIMEUP,
+		),
+		SG.Text(
+			key="_task2DownTimer_",
+			**TEXTADJTIMEDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="U2M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="U2S-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D2M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D2S-",
+			**ADJBTNDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="D2B",
+			**BTNBOWLSDOWN,
+		),
+		SG.Text(
+			key="_task2BowlsCount_",
+			**TEXTBOWLSCOUNT,
+		),
+		SG.Btn(
+			key="U2B",
+			**BTNBOWLSUP,
+		),
+	],
+]
+
+CLMTASK3 = [
+	[
+		SG.Spin(
+			initial_value=currentData[TASK3NAME],
+			key="_task3_spin_",
+			**TASKSPINBOX,
+		),
+	],
+	[
+		SG.Button(
+			key="task3+",
+			**BTNTASKUP,
+		),
+		SG.Btn(
+			key="task3+nt",
+			**BTNUPNOTIME,
+		),
+	],
+	[
+		SG.Text(
+			key="_task3count_",
+			**TEXTTASKCOUNT,
+		),
+	],
+	[
+		SG.Button(
+			key="task3-",
+			**BTNTASKDOWN,
+		),
+		SG.Btn(
+			key="task3reset",
+			**BTNRESETONE,
+		),
+	],
+	[
+		SG.Btn(
+			key="U3M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="U3S+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D3M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D3S+",
+			**ADJBTNUP,
+		),
+	],
+	[
+		SG.Text(
+			key="_task3UpTimer_",
+			**TEXTADJTIMEUP,
+		),
+		SG.Text(
+			key="_task3DownTimer_",
+			**TEXTADJTIMEDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="U3M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="U3S-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D3M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D3S-",
+			**ADJBTNDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="D3B",
+			**BTNBOWLSDOWN,
+		),
+		SG.Text(
+			key="_task3BowlsCount_",
+			**TEXTBOWLSCOUNT,
+		),
+		SG.Btn(
+			key="U3B",
+			**BTNBOWLSUP,
+		),
+	],
+]
+
+CLMTASK4 = [
+	[
+		SG.Spin(
+			initial_value=currentData[TASK4NAME],
+			key="_task4_spin_",
+			**TASKSPINBOX,
+		),
+	],
+	[
+		SG.Button(
+			key="task4+",
+			**BTNTASKUP,
+		),
+		SG.Btn(
+			key="task4+nt",
+			**BTNUPNOTIME,
+		),
+	],
+	[
+		SG.Text(
+			key="_task4count_",
+			**TEXTTASKCOUNT,
+		),
+	],
+	[
+		SG.Button(
+			key="task4-",
+			**BTNTASKDOWN,
+		),
+		SG.Btn(
+			key="task4reset",
+			**BTNRESETONE,
+		),
+	],
+	[
+		SG.Btn(
+			key="U4M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="U4S+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D4M+",
+			**ADJBTNUP,
+		),
+		SG.Btn(
+			key="D4S+",
+			**ADJBTNUP,
+		),
+	],
+	[
+		SG.Text(
+			key="_task4UpTimer_",
+			**TEXTADJTIMEUP,
+		),
+		SG.Text(
+			key="_task4DownTimer_",
+			**TEXTADJTIMEDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="U4M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="U4S-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D4M-",
+			**ADJBTNDOWN,
+		),
+		SG.Btn(
+			key="D4S-",
+			**ADJBTNDOWN,
+		),
+	],
+	[
+		SG.Btn(
+			key="D4B",
+			**BTNBOWLSDOWN,
+		),
+		SG.Text(
+			key="_task4BowlsCount_",
+			**TEXTBOWLSCOUNT,
+		),
+		SG.Btn(
+			key="U4B",
+			**BTNBOWLSUP,
+		),
+	],
+]
+
+CLMTIMER = [
+	[
+		SG.Text(
+			"000:00",
+			size=(6, 1),
+			text_color=TIMERUPTEXTCOLOR,
+			font=(FONT, TIMERFONTSZ),
+			justification="center",
+			key="_timer_",
+		),
+	],
+]
+
+mainLayout = [
+	[
+		SG.Col(CLMTIMER),
+	],
+	[
+		SG.Text(
+			key="_upTime_",
+			**TEXTUPTIME,
+		),
+		SG.Text(
+			key="_downTime_",
+			**TEXTDOWNTIME,
+		),
+		SG.Col(
+			[
+				[
+					SG.Btn(
+						key="Quit",
+						**BTNQUIT,
+					),
+				],
+				[
+					SG.Btn(
+						key="zeroAll",
+						**BTNZEROALL,
+					),
+				],
+			],
+		),
+	],
+	[
+		SG.Col(CLMTASK1),
+		SG.Col(CLMTASK2),
+		SG.Col(CLMTASK3),
+		SG.Col(CLMTASK4),
+	],
+]
+
+mainWindowParms = {
+	"title": MAINWINDOWTITLE,
+	"layout": mainLayout,
+	"location": (0, 0),
+	"element_padding": (0, 0),
+	"no_titlebar": False,
+}
+
+mainFrame = SG.Window(
+	**mainWindowParms
+).finalize()
+
+
+def nowStr(dtObj=DT.now()):
+	return dtObj.strftime("%Y%m%d.%H%M%S")
+
+
+def setTimer():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	ticks_ = ticks // ticksPerSecond
+	mins_ = int(ticks_ // 60)
+	secs_ = int(ticks_ % 60)
+	timeStr = f"{mins_:03d}:{secs_:02d}"
+	if directionUp is True and timerRunning is True:
+		mainFrame.Element("_upTime_").Update(value=(makeTime(currentData[UPSEC] - ticks_)))
+		mainFrame.Element("_downTime_").Update(value=(makeTime(0)))
+		mainFrame.Element("_timer_").Update(value=timeStr)
+	elif directionUp is False and timerRunning is True:
+		mainFrame.Element("_upTime_").Update(value=(makeTime(0)))
+		mainFrame.Element("_downTime_").Update(value=(makeTime(currentData[DOWNSEC] - ticks_)))
+		mainFrame.Element("_timer_").Update(value=timeStr)
+	else:
+		mainFrame.Element("_upTime_").Update(value=(makeTime(0)))
+		mainFrame.Element("_downTime_").Update(value=(makeTime(0)))
+		mainFrame.Element("_timer_").Update(value=timeStr)
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def updateAll():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	# update timer and cycleCount
+	tempTimerVAL = ticks // ticksPerSecond
+	setTimer()
+	mainFrame.Element("_task1count_").Update(value=(f"{currentData[TASK1COUNT]:03d}"))
+	mainFrame.Element("_task2count_").Update(value=(f"{currentData[TASK2COUNT]:03d}"))
+	mainFrame.Element("_task3count_").Update(value=(f"{currentData[TASK3COUNT]:03d}"))
+	mainFrame.Element("_task4count_").Update(value=(f"{currentData[TASK4COUNT]:03d}"))
+
+	mainFrame.Element("_task1BowlsCount_").Update(value=(f"{currentData[TASK1BOWLSCOUNT]:02d}"))
+	mainFrame.Element("_task2BowlsCount_").Update(value=(f"{currentData[TASK2BOWLSCOUNT]:02d}"))
+	mainFrame.Element("_task3BowlsCount_").Update(value=(f"{currentData[TASK3BOWLSCOUNT]:02d}"))
+	mainFrame.Element("_task4BowlsCount_").Update(value=(f"{currentData[TASK4BOWLSCOUNT]:02d}"))
+
+	TUpTimer = makeTime(currentData[TASK1UPTIMER])
+	TDownTimer = makeTime(currentData[TASK1DOWNTIMER])
+	mainFrame.Element("_task1UpTimer_").Update(value=f"{TUpTimer}")
+	mainFrame.Element("_task1DownTimer_").Update(value=f"{TDownTimer}")
+
+	TUpTimer = makeTime(currentData[TASK2UPTIMER])
+	TDownTimer = makeTime(currentData[TASK2DOWNTIMER])
+	mainFrame.Element("_task2UpTimer_").Update(value=f"{TUpTimer}")
+	mainFrame.Element("_task2DownTimer_").Update(value=f"{TDownTimer}")
+
+	TUpTimer = makeTime(currentData[TASK3UPTIMER])
+	TDownTimer = makeTime(currentData[TASK3DOWNTIMER])
+	mainFrame.Element("_task3UpTimer_").Update(value=f"{TUpTimer}")
+	mainFrame.Element("_task3DownTimer_").Update(value=f"{TDownTimer}")
+
+	TUpTimer = makeTime(currentData[TASK4UPTIMER])
+	TDownTimer = makeTime(currentData[TASK4DOWNTIMER])
+	mainFrame.Element("_task4UpTimer_").Update(value=f"{TUpTimer}")
+	mainFrame.Element("_task4DownTimer_").Update(value=f"{TDownTimer}")
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def incCount(_inCount_):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	_T1_ = _inCount_ + 1
+	if _T1_ > 999:
+		_T1_ = 999
+	return _T1_
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def incBowlCount(_inCount_):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	_T1_ = _inCount_ + 1
+	if _T1_ > 99:
+		_T1_ = 99
+	return _T1_
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def decCount(_inCount_):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	TI = _inCount_ - 1
+	if TI < 0:
+		TI = 0
+	return TI
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def decBowlCount(_inCount_):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	_T1_ = _inCount_ - 1
+	if _T1_ < 0:
+		_T1_ = 0
+	return _T1_
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def incTime(inSeconds, increment):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	TInSeconds = inSeconds + increment
+	TMin = TInSeconds // 60
+	# TSec = TInSeconds % 60
+	if TMin > 999:
+		TInSeconds = (999 * 60) + 59
+	return TInSeconds
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def decTime(inSeconds, increment):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	TInSeconds = inSeconds - increment
+	if TInSeconds < 0:
+		TInSeconds = 0
+	return TInSeconds
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def updateMainBackground(COLOR):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	# put change background code
+	mainFrame.Element("_timer_").Update(background_color=COLOR)
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def setTimerOff():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global ticks, directionUp
+	directionUp = True
+	ticks = 0
+	mainFrame.Element("_timer_").Update(**TEXTTIMEROFFPARMS)
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def setTimerUp():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global directionUp, ticks
+	directionUp = True
+	ticks = 0
+	mainFrame.Element("_timer_").Update(**TEXTTIMERUPPARMS)
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def zeroStuff():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global ticks, directionUp, currentData, timerRunning
+	ticks = 0
+	timerRunning = False
+	directionUp = True
+	currentData[EVENTS] = []
+	currentData[TASK1COUNT] = 0
+	currentData[TASK2COUNT] = 0
+	currentData[TASK3COUNT] = 0
+	currentData[TASK4COUNT] = 0
+	currentData[TASK1BOWLSCOUNT] = 0
+	currentData[TASK2BOWLSCOUNT] = 0
+	currentData[TASK3BOWLSCOUNT] = 0
+	currentData[TASK4BOWLSCOUNT] = 0
+	pickleIt(currentData[FILENAME], currentData)
+	updateAll()
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def startTimer():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global timerRunning, currentData, ticks, directionUp
+	timerRunning = True
+	ticks = 0
+	directionUp = True
+	mainFrame.Element("_timer_").Update(**TEXTTIMERUPPARMS)
+	setTimer()
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def stopTimer():
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global timerRunning, ticks
+	timerRunning = False
+	setTimerOff()
+	updateAll()
+	mainFrame.Element("_timer_").Update(**TEXTTIMEROFFPARMS)
+	ticks = 0
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def getValues(inValues):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global currentData
+	for key, item in DEFAULTSVALUESNDX.items():
+		currentData[key] = inValues[item]
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def addEvent(event2add):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global currentData
+	return
+	entryToAdd = [nowStr(DT.now()), event2add]
+	# currentData[EVENTS].append(entryToAdd)
+	pickleIt(currentData[FILENAME], currentData)
+	with open(CONFIGDIRECTORY + currentData[TEXTNAME], "ta") as FDOut:
+		outStr = ""
+		outStr += f"""{entryToAdd}	{currentData[TASK1COUNT]}	{currentData[TASK2COUNT]}	{currentData[TASK3COUNT]}	{currentData[TASK4COUNT]}
+"""
+		# print(outStr)
+		FDOut.writelines(outStr)
+		FDOut.flush()
+		FDOut.close()
+	# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+
+def doEvents(event):
+	# fold here ⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱⟱
+	global currentData
+	if timerRunning is True:
+		return
+	if event == "zeroAll":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		zeroStuff()
+		stopTimer()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "task1+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1COUNT] = incCount(currentData[TASK1COUNT])
+		currentData[UPSEC] = currentData[TASK1UPTIMER]
+		currentData[DOWNSEC] = currentData[TASK1DOWNTIMER]
+		updateAll()
+		startTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task2+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2COUNT] = incCount(currentData[TASK2COUNT])
+		currentData[UPSEC] = currentData[TASK2UPTIMER]
+		currentData[DOWNSEC] = currentData[TASK2DOWNTIMER]
+		updateAll()
+		startTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task3+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3COUNT] = incCount(currentData[TASK3COUNT])
+		currentData[UPSEC] = currentData[TASK3UPTIMER]
+		currentData[DOWNSEC] = currentData[TASK3DOWNTIMER]
+		updateAll()
+		startTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task4+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4COUNT] = incCount(currentData[TASK4COUNT])
+		currentData[UPSEC] = currentData[TASK4UPTIMER]
+		currentData[DOWNSEC] = currentData[TASK4DOWNTIMER]
+		updateAll()
+		startTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "task1-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1COUNT] = decCount(currentData[TASK1COUNT])
+		updateAll()
+		stopTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task2-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2COUNT] = decCount(currentData[TASK2COUNT])
+		updateAll()
+		stopTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task3-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3COUNT] = decCount(currentData[TASK3COUNT])
+		updateAll()
+		stopTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task4-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4COUNT] = decCount(currentData[TASK4COUNT])
+		updateAll()
+		stopTimer()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U1M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1UPTIMER] = incTime(currentData[TASK1UPTIMER], 60)
+		updateAll()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U1S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1UPTIMER] = incTime(currentData[TASK1UPTIMER], 1)
+		updateAll()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D1M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1DOWNTIMER] = incTime(currentData[TASK1DOWNTIMER], 60)
+		updateAll()
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D1S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1DOWNTIMER] = incTime(currentData[TASK1DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U2M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2UPTIMER] = incTime(currentData[TASK2UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U2S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2UPTIMER] = incTime(currentData[TASK2UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D2M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2DOWNTIMER] = incTime(currentData[TASK2DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D2S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2DOWNTIMER] = incTime(currentData[TASK2DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U3M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3UPTIMER] = incTime(currentData[TASK3UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U3S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3UPTIMER] = incTime(currentData[TASK3UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D3M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3DOWNTIMER] = incTime(currentData[TASK3DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D3S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3DOWNTIMER] = incTime(currentData[TASK3DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U4M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4UPTIMER] = incTime(currentData[TASK4UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U4S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4UPTIMER] = incTime(currentData[TASK4UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D4M+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4DOWNTIMER] = incTime(currentData[TASK4DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D4S+":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4DOWNTIMER] = incTime(currentData[TASK4DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U1M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1UPTIMER] = decTime(currentData[TASK1UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U1S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1UPTIMER] = decTime(currentData[TASK1UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D1M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1DOWNTIMER] = decTime(currentData[TASK1DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D1S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1DOWNTIMER] = decTime(currentData[TASK1DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U2M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2UPTIMER] = decTime(currentData[TASK2UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U2S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2UPTIMER] = decTime(currentData[TASK2UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D2M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2DOWNTIMER] = decTime(currentData[TASK2DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D2S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2DOWNTIMER] = decTime(currentData[TASK2DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U3M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3UPTIMER] = decTime(currentData[TASK3UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U3S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3UPTIMER] = decTime(currentData[TASK3UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D3M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3DOWNTIMER] = decTime(currentData[TASK3DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D3S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3DOWNTIMER] = decTime(currentData[TASK3DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U4M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4UPTIMER] = decTime(currentData[TASK4UPTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U4S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4UPTIMER] = decTime(currentData[TASK4UPTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D4M-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4DOWNTIMER] = decTime(currentData[TASK4DOWNTIMER], 60)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D4S-":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4DOWNTIMER] = decTime(currentData[TASK4DOWNTIMER], 1)
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "task1+nt":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1COUNT] = incCount(currentData[TASK1COUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task2+nt":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2COUNT] = incCount(currentData[TASK2COUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task3+nt":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3COUNT] = incCount(currentData[TASK3COUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task4+nt":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4COUNT] = incCount(currentData[TASK4COUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "task1reset":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1COUNT] = 0
+		currentData[TASK1BOWLSCOUNT] = 0
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task2reset":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2COUNT] = 0
+		currentData[TASK2BOWLSCOUNT] = 0
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task3reset":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3COUNT] = 0
+		currentData[TASK3BOWLSCOUNT] = 0
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "task4reset":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4COUNT] = 0
+		currentData[TASK4BOWLSCOUNT] = 0
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "D1B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1BOWLSCOUNT] = decBowlCount(currentData[TASK1BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D2B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2BOWLSCOUNT] = decBowlCount(currentData[TASK2BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D3B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3BOWLSCOUNT] = decBowlCount(currentData[TASK3BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "D4B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4BOWLSCOUNT] = decBowlCount(currentData[TASK4BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+
+	elif event == "U1B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK1BOWLSCOUNT] = incBowlCount(currentData[TASK1BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U2B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK2BOWLSCOUNT] = incBowlCount(currentData[TASK2BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U3B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK3BOWLSCOUNT] = incBowlCount(currentData[TASK3BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "U4B":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		currentData[TASK4BOWLSCOUNT] = incBowlCount(currentData[TASK4BOWLSCOUNT])
+		updateAll()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+# fold here ⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰⟰
+
+getData(currentData[FILENAME])
+updateAll()
+stopTimer()
+while True:  # Event Loop
+	event, values = mainFrame.Read(timeout=interruptRate)  # use as high of a timeout value as you can
+	# 	print(f"""
+	# event |{event}|
+	# values |{values}|
+	# """)
+	if (event is None) or (event == "Quit"):  # X or quit button clicked
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		addEvent(event)
+		stopTimer()
+		pickleIt(currentData[FILENAME], currentData)
+		break
+	# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	elif event == "__TIMEOUT__":
+		# fold here ⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥⥥
+		if directionUp is True:
+			ticks += 1
+		else:
+			ticks -= 1
+		if abs(ticks) > MAXTICKS:
+			ticks = 0
+		# print(f"ticks {ticks}")
+		if timerRunning is True:
+			upTicks = int(currentData[UPSEC] * ticksPerSecond)
+			downTicks = int(currentData[DOWNSEC] * ticksPerSecond)
+			if (directionUp is True) and (ticks > upTicks):
+				mainFrame.Element("_timer_").Update(**TEXTTIMERDOWNPARMS)
+				directionUp = False
+				ticks = downTicks
+			elif (directionUp is False) & (ticks <= ticksPerSecond):
+				stopTimer()
+		setTimer()
+		# fold here ⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣⥣
+	else:
+		doEvents(event)
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# end of while True
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+
+
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
+# end of biditi.proper
+# #*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#*#
