+++ newFM.py
@@ -255,6 +255,15 @@ FMAXPSG_SCTN0905_TUPDICTSTRADD = "FMAXPSG_SCTN0905_TUPDICTSTRADD"  # add a str t
 FMAXPSG_SCTN0905_TUPDICTSTRSTRADD = "FMAXPSG_SCTN0905_TUPDICTSTRSTRADD"  # add a str to a tupdict <NAC><TUPDICTNAME><STR><STR>
 FMAXPSG_SCTN0905_TUPDICTSTRVALADD = "FMAXPSG_SCTN0905_TUPDICTSTRVALADD"  # add a value to a tupdict <NAC><TUPDICTNAME><STR><VAL>
 FMAXPSG_SCTN0905_TUPDICTVALADD = "FMAXPSG_SCTN0905_TUPDICTVALADD"  # add a value to a tupdict <NAC><TUPDICTNAME><KEY><VAL>
+FMAXPSG_SCTN0906_BTN_STRADD = "FMAXPSG_SCTN0906_BTN_STRADD"  # add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL>
+FMAXPSG_SCTN0906_BTN_TUPVALVALADD = "FMAXPSG_SCTN0906_BTN_TUPVALVALADD"  # add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL1><VAL2>
+FMAXPSG_SCTN0906_BTN_VALADD = "FMAXPSG_SCTN0906_BTN_VALADD"  # add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL>
+FMAXPSG_SCTN0906_BTNDEF = "FMAXPSG_SCTN0906_BTNDEF"  # define a button <NAC><BTNNAME>
+FMAXPSG_SCTN0907_SPIN_DICTSTRADD = "FMAXPSG_SCTN0907_SPIN_DICTSTRADD"  # add a VAL to the SPINDICT <NAC><SPINNAME><KEY><VAL>
+FMAXPSG_SCTN0907_SPIN_DICTTUPVALVALADD = "FMAXPSG_SCTN0907_SPIN_DICTTUPVALVALADD"  # add a (VAL1, VAL2) to the SPINDICT <NAC><SPINNAME><KEY><VAL1><VAL2>
+FMAXPSG_SCTN0907_SPIN_VALUESSTRADD = "FMAXPSG_SCTN0907_SPIN_VALUESSTRADD"  # add a STR to the values list <NAC><SPINNAME><STR>
+FMAXPSG_SCTN0907_SPIN_VALUESVALADD = "FMAXPSG_SCTN0907_SPIN_VALUESVALADD"  # add a VAL to the values list <NAC><SPINNAME><VAL>
+FMAXPSG_SCTN0907_SPINDEF = "FMAXPSG_SCTN0907_SPINDEF"  # define a spin box entry <NAC><SPINNAME>
 
 
 FMAXFM_AXLST = [
@@ -308,6 +317,15 @@ FMAXFM_AXLST = [
 	FMAXPSG_SCTN0905_TUPDICTSTRSTRADD,  # add a str to a tupdict <NAC><TUPDICTNAME><STR><STR>
 	FMAXPSG_SCTN0905_TUPDICTSTRVALADD,  # add a value to a tupdict <NAC><TUPDICTNAME><STR><VAL>
 	FMAXPSG_SCTN0905_TUPDICTVALADD,  # add a value to a tupdict <NAC><TUPDICTNAME><KEY><VAL>
+	FMAXPSG_SCTN0906_BTN_STRADD,  # add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL>
+	FMAXPSG_SCTN0906_BTN_TUPVALVALADD,  # add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL1><VAL2>
+	FMAXPSG_SCTN0906_BTN_VALADD,  # add a (VAL,VAL) to a tupdict <NAC><BTNNAME><KEY><VAL>
+	FMAXPSG_SCTN0906_BTNDEF,  # define a button <NAC><BTNNAME>
+	FMAXPSG_SCTN0907_SPIN_DICTSTRADD,  # add a VAL to the SPINDICT <NAC><SPINNAME><KEY><VAL>
+	FMAXPSG_SCTN0907_SPIN_DICTTUPVALVALADD,  # add a (VAL1, VAL2) to the SPINDICT <NAC><SPINNAME><KEY><VAL1><VAL2>
+	FMAXPSG_SCTN0907_SPIN_VALUESSTRADD,  # add a STR to the values list <NAC><SPINNAME><STR>
+	FMAXPSG_SCTN0907_SPIN_VALUESVALADD,  # add a VAL to the values list <NAC><SPINNAME><VAL>
+	FMAXPSG_SCTN0907_SPINDEF,  # define a spin box entry <NAC><SPINNAME>
 ]
